### 2024-04-16 23072ca
Included JOSS badge
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
