### 2024-04-30 #6628 @srmnitc
👋, @ptmerz @corettialessandro @liamhuber would any of you have time and be willing to review this submission for JOSS? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Thanks for your time!
