### 2024-05-01 #6628 @sntioudis
Hi @liamhuber, thank you for taking the time to review the PAPRECA documentation pages!

> @sntioudis what are the odds this is an overly-restrictive statement? I'm running OSX on my laptop, so I have ready access to unix but not linux. My other machine runs Windows. If you think there's a chance unix will cut it, I'll give it a go and see what happens.

Unfortunately, we have no access to macOS and therefore, we did not test PAPRECA on such a system. PAPRECA was only tested on Linux and this is why this overly restrictive statement is present in the documentation: _"The current version of PAPRECA (1.0) runs only on LINUX-based systems. Cross-platform compatibility will be available in future versions. At the moment, non-LINUX users can run PAPRECA on a virtual machine (e.g., [VirtualBox](https://www.virtualbox.org/) or Windows Subsystem for Linux)."_

That being said, there is a chance that PAPRECA can be installed on macOS but there is no way for us to check if it is running properly. Also, as we have no access to macOS we will not be able to support you and make the necessary changes if any issues manifest during the review.

> I'd prefer to avoid setting up a VM or dual partition on my windows machine before even starting the review.

Finally, I just wanted to mention that if you are a Windows user there is not really a need to set up a virtual machine or a partition in order to run PAPRECA. You can use the [Windows subsystem for Linux (WSL)](https://learn.microsoft.com/en-us/windows/wsl/install) which gives you a Linux terminal on Windows. Installing WSL can be as easy as running a single command from your PowerShell.
