### 2024-05-01 #6628 @srmnitc
> From the [review guidelines](https://joss.readthedocs.io/en/latest/review_checklist.html#functionality) I am inferring that I should actually take this thing for a spin. My first step was to check out the installation process, andthese were [easy to find](https://sntioudis.github.io/papreca/md_installation.html) and docs overall look well organized and readable (kudos to the author team so far!)
>
> I got to here:
>
> > The current version of PAPRECA (1.0) runs only on LINUX-based systems. Cross-platform compatibility will be available in future versions. At the moment, non-LINUX users can run PAPRECA on a virtual machine (e.g., [VirtualBox](https://www.virtualbox.org/) or Windows Subsystem for Linux).
>
> @sntioudis what are the odds this is an overly-restrictive statement? I'm running OSX on my laptop, so I have ready access to _unix_ but not _linux_. My other machine runs Windows. If you think there's a chance unix will cut it, I'll give it a go and see what happens.
>
> If linux is a hard requirement, and assuming one of the other people tagged is able to review, I'll recuse myself by reason of laziness. As long as I'm not critically required, I'd prefer to avoid setting up a VM or dual partition on my windows machine before even starting the review.

@liamhuber If you would still like to go ahead with a review, let us know. For the review itself, I will start a separate issue. We will start the review once we have found two reviewers. At the moment, a confirmation from you would be enough :)
