### 2024-05-02 #6628 @srmnitc
👋 @ptmerz @corettialessandro @liamhuber would any of you have time and be willing to review this submission for JOSS? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Thanks for your time!

> @sntioudis thanks for the WSL tip! @srmnitc, go ahead and ~invite~ tag/assign me, I'll start working on it next week.

Thanks a lot @liamhuber, once I find one more reviewer I will start review issue and close this one.
