### 2024-05-02 #6714 @editorialbot
[REVIEW]: PAPRECA: A parallel hybrid off-lattice kinetic Monte Carlo/molecular dynamics simulator
**Submitting author:** @sntioudis (<a href="http://orcid.org/0009-0000-8095-1727">Stavros Ntioudis</a>)
**Repository:** https://github.com/sntioudis/papreca
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
