### 2024-05-09 e347443
Update installation.md
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
