### 2024-05-09 sntioudis/papreca#2 @liamhuber
Release management
Distribution via clone-and-build A-OK, but I like how lammps manages their versioning for this process, and recommend following a similar pattern. Namely, you can already make your first release tag (in contrast to lammps, I would strongly recommend using [semantic versioning](https://semver.org/) for such tags), then make a release branch and park it at that tag. Then in the installation docs add the same snippet you have for lammps -- `git checkout release; git pull`

This is not a huge deal, but it will make it easy for you to have stable (release) and bleeding-edge (main) versions right out of the gate.

This is a non-blocking issue coming from https://github.com/openjournals/joss-reviews/issues/6714
