+++ README.md
@@ -12,7 +12,7 @@ Open-source and community guidelines
 -------------
 PAPRECA is an open-source software that is distributed under the terms of the [GNU General Public License, version 2](https://en.wikipedia.org/wiki/GNU_General_Public_License). PAPRECA is designed as an easy-to-modify and easy-to-extend software. For example, new predefined event templates can be added to the original code.
 
-In the spirit of open source, we encourage contributions from other Authors! Kindly use pull requests to contribute to the code after ensuring your local fork passes ALL tests. Furthermore, you can use GitHub issues for other forms of contribution (e.g., bug fixes, code optimization, new examples, corrections to documentation).
+In the spirit of open source, we encourage contributions from other Authors! Kindly use pull requests to contribute to the code after ensuring your local fork passes all tests. Furthermore, you can use GitHub issues for other forms of contribution (e.g., bug fixes, code optimization, new examples, corrections to documentation). Note that, development will solely be performed on the main branch. Merges with the release branch will only happen after ensuring that the code on the main branch is ready for production. Appropriate (semantic) versioning tags will be used to link the released version against the relevant commit.
 
 Additionally, enquiries can be directed to the original developer of this software, Stavros Ntioudis, via e-mail: [stavros.ntioudis20@imperial.ac.uk](mailto:stavros.ntioudis20@imperial.ac.uk).
 
