### 2024-05-10 sntioudis/papreca#2 @sntioudis
This is a good point. I created a new (release) branch to address this and made the relevant changes to the documentation.
