### 2024-05-15 #6714 @liamhuber
Write-up comment:

> Overall, kMC techniques are, on one hand, less accurate, but on the other hand, more
computationally efficient than other atomistic techniques such as the density functional theory (DFT) or molecular dynamics (MD).

I find the comparison to DFT to be apples-to-oranges, it's equivalent to saying "more computationally efficient that MEAM potentials". DFT is a tool for transforming structures (position+chemistry) into energies and forces (and electron density); kMC is a tool for evolving structures in time. In contrast, the comparison to MD is apples-to-apples. Indeed, one could use DFT (assuming arbitrarily powerful compute) and MD and kMC all together! So the real contrast here is kMC vs brute-force MD for temporal evolution.

IMO the comparison to brute-force MD is already sufficient and DFT should simply be removed. If you really want multiple items to contrast, you could consider mentioning some other acceleration technique, e.g. stuff by Danny Perez and Arthur Voter comes to mind.
