### 2024-05-15 sntioudis/papreca#6 @liamhuber
Tests: Parallelize by change steps to separate jobs
A non-blocking issue for https://github.com/openjournals/joss-reviews/issues/6714

My understanding of the test flow is this: `run_ALLtests.yml` has a series of steps that creates the environment, builds lammps and papreca, and then three steps for actual tests managed by bash scripts; the middle of these is run across five different (but fixed) seeds, and in turn is a script running a variety of sub-tests all following the pattern: build input, run python to execute papreca and analyze the results.

There is no game-stopping problem here, but there are two inefficiencies:

The first is that the functionality tests are all run in serial. I see in [the test logs](https://github.com/sntioudis/papreca/actions/runs/8628715594/job/23651395721) that actually _running_ papreca takes insignificant time (<2m) compared to building the dependencies (>8m for lammps), so in practice there is no big gain by parallelizing them, _but_ doing so does free you to make the actual runs more expensive without making the tests take so much longer.

The second is informational; currently if the tests fail in `Run all event tests for 5 different random seeds` the dashboard will just show you that they failed somewhere in some event for some seed, and you'll need to dig into the log to get any information beyond that.

The alternative structure would be to break the run from a single job with all these steps into a variety of jobs, e.g. setup, source tests, a [matrix](https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs#using-a-matrix-strategy) (across seed numbers and event type shell scripts (basically breaking apart test_all_events.sh)) for the event tests, and the hybrid test.

Of course once these are all separate jobs they'll attempt to run concurrently, which is just what we want for parallel speedup but we don't want the tests to run until the setup is done! You can force all the test jobs wait for the setup using [needs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idneeds).

Further, they are really separate jobs on separate VMs, so in order to make sure the papreca binary (and anything else you need?) available you will need to store it as an [artifact](https://docs.github.com/en/actions/using-workflows/storing-workflow-data-as-artifacts#about-workflow-artifacts).

Since the event tests -- which would most heavily benefit from parallelization -- are not the chokepoint, the hard cost-benefit here is low. There is a soft benefit of learning about github actions and commercial cloud compute management, but this is just a pointer for a direction you _might_ go, and not a fundamental problem with the repo.
