### 2024-05-15 sntioudis/papreca#7 @liamhuber
Tests: cache lammps?
A non-blocking issue for https://github.com/openjournals/joss-reviews/issues/6714

Since [half your test time](https://github.com/sntioudis/papreca/actions/runs/8628715594/job/23651395721) is going to building lammps, you may wish to consider [caching the built executable](https://github.com/actions/cache). Just remember to include the commit hash at the head of the [`release` branch you're cloning](https://github.com/sntioudis/papreca/blob/0f10dc3625b5342d8c8ae7e38b567baa40a1231a/.github/workflows/run_ALLtests.yml#L44) in the cache key -- then if `release` has been updated the key will be new and you can fail your cache hit and correctly re-build lammps.

Disclaimer: I haven't ever tried caching a built executable. I hope it will be as easy as caching the compiled binary, and I'd start by simply trying that, but it's possible some linked libraries will sink you. As long as you're accessing the cache on the same runner (OS and OS version), I am cautiously optimistic it will work OK.
