### 2024-05-15 sntioudis/papreca#8 @liamhuber
Introduce a lower-bound for lammps?
A non-blocking issue for https://github.com/openjournals/joss-reviews/issues/6714

Right now your installation instructions simply say to target `release`. It might be nice to target a (or multiple?) older versions of lammps and explicitly test against them.

This would be super straightforward by reformulating your existing `run_ALLtests.yml` workflow as a [reusable workflow](https://docs.github.com/en/actions/using-workflows/reusing-workflows) that takes the lammps branch as input. The current tests would then just be the special case of that script targeting `release`.
