### 2024-05-15 sntioudis/papreca#9 @liamhuber
Don't let lammps break papreca
Not ever for stable releases, and not without learning about it quickly for the bleeding-edge.

A **blocking** issue for https://github.com/openjournals/joss-reviews/issues/6714

Your install procedure currently targets the lammps `release` branch. This branch is subject to change, and may introduce something that breaks papreca. Two things should be done to mitigate this:

- [x] The docs for stable releases (and any build tools you introduce in the future, e.g. as part of releasing on [conda](https://github.com/sntioudis/papreca/issues/4) should make sure they target a _specific_ lammps release which is known with certainty to work. Such stable-version-docs can still talk about targeting `release`, but should be explicit about which version is guaranteed to work. (Targeting a particular version for release like this synergizes well with the steps for [testing some lower bound(s)](https://github.com/sntioudis/papreca/issues/8).)
- [x] Tests targeting the `release` should be run on a regular [schedule](https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#schedule) (weekly or even daily) and not just [on manual trigger](https://github.com/sntioudis/papreca/blob/0f10dc3625b5342d8c8ae7e38b567baa40a1231a/.github/workflows/run_ALLtests.yml#L13), and you should make sure your watch settings for the CI are set so you'll get the alert -- e.g. by e-mail if the project maintainer is not on github regularly enough to be sure to see it there. IMO it's ok to have "bleeding edge" advice that simply sends people to lammps' `release` branch, but you still want to find out promptly if that advice goes sour.
