### 2024-05-15 sntioudis/papreca#10 @liamhuber
Tests: fail on physics failure
An issue for https://github.com/openjournals/joss-reviews/issues/6714 (I'm on the fence whether it should be blocking)

I am very happy that the event types are [all](https://github.com/sntioudis/papreca/blob/main/tests/event%20tests/test_all_events.sh) getting [tested](https://github.com/sntioudis/papreca/blob/0f10dc3625b5342d8c8ae7e38b567baa40a1231a/.github/workflows/run_ALLtests.yml#L76). If papreca actually fails for any of them, you nicely get a bad exit ([e.g.](https://github.com/sntioudis/papreca/blob/0f10dc3625b5342d8c8ae7e38b567baa40a1231a/tests/event%20tests/formation%20events/test_formations.py#L101)). Further, the various `compareArraysAndPrintStats` nicely give some human-readable report on actual metrics from the run. However, these metrics can be absolutely garbage and the test will still pass; the test itself should be made to fail if the printed metrics would concern a human reader.
