### 2024-05-15 sntioudis/papreca#11 @liamhuber
Tests: on float comparison
An issue for https://github.com/openjournals/joss-reviews/issues/6714

Throughout the [event tests](https://github.com/sntioudis/papreca/tree/main/tests/event%20tests) float comparisons are consistently made with `==`. This won't give reliable results, so, since you already have numpy as a dependency, these should be replaced with `numpy.isclose`. As a bunch of your data is `np.array` already, and is-close works on arrays, this may allow you to slim down your line count at the same time.
