### 2024-05-16 e0a5929
Update paper.md (removed DFT from comparison with kMC)
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
