+++ paper/paper.md
@@ -37,7 +37,7 @@ Kinetic Monte Carlo (kMC) is an atomistic and stochastic simulation technique th
 
 # Statement of need
 
-KMC models have been deployed to investigate non-equilibrium dynamics and properties of thin films [@Ntioudis2023Oct], nanoparticles [@Turner2016Apr], quantum dots [@Zhu2007May], semiconductors [@vanderKaap2016Feb], catalysts [@Stamatakis2012Dec], energy-storage devices [@Abbott2022Feb], interstellar grain chemistry [@Cuppen2013Dec], protein folding [@Makarov2001Jun], and enzyme reactions [@Slepoy2008May]. Overall, kMC techniques are, on one hand, less accurate, but on the other hand, more computationally efficient than other atomistic techniques such as the density functional theory (DFT) or molecular dynamics (MD). This is justified by the fact that kMC does not describe atomic vibrations explicitly but evolves the system through discrete elementary processes (e.g., diffusions, depositions, reactions etc.) [@Gillespie1976Dec ; @Fichthorn1991Jul]. In any case, the efficiency of kMC models unlocks the possibility for long-timescale simulations with molecular-level resolution beyond the timescales accessible to MD. Note that efficient MD algorithms can be used to simulate systems with $10^{4}$-$10^{6}$ atoms from ps to $\mu$s [@Thompson2022Feb]).
+KMC models have been deployed to investigate non-equilibrium dynamics and properties of thin films [@Ntioudis2023Oct], nanoparticles [@Turner2016Apr], quantum dots [@Zhu2007May], semiconductors [@vanderKaap2016Feb], catalysts [@Stamatakis2012Dec], energy-storage devices [@Abbott2022Feb], interstellar grain chemistry [@Cuppen2013Dec], protein folding [@Makarov2001Jun], and enzyme reactions [@Slepoy2008May]. Overall, kMC techniques are, on one hand, less accurate, but on the other hand, more computationally efficient than molecular dynamics (MD). This is justified by the fact that kMC does not describe atomic vibrations explicitly but evolves the system through discrete elementary processes (e.g., diffusions, depositions, reactions etc.) [@Gillespie1976Dec ; @Fichthorn1991Jul]. In any case, the efficiency of kMC models unlocks the possibility for long-timescale simulations with molecular-level resolution beyond the timescales accessible to MD. Note that efficient MD algorithms can be used to simulate systems with $10^{4}$-$10^{6}$ atoms from ps to $\mu$s [@Thompson2022Feb]).
 
 Typically, on-lattice kMC models select atomistic events from a predefined table and execute them on fixed lattice sites [@Andersen2019Apr]. The use of fixed lattice sites contributes to the computational efficiency of on-lattice algorithms but introduces obstacles associated with the study of unordered materials.
 Several lattice-based open-source kMC software are available in the bibliography. Examples include the KMCLib [@Leetmaa2014Sep], lattice_mc [@Morgan2017May], KMC_Lattice v2.0 [@Heiber2019Jan], Excimontec v1.0 [@Heiber2020Sep], MonteCoffee [@Jorgensen2018Sep], and KIMERA [@Martin2020Sep]. Additionally, the Stochastic Parallel PARticle Kinetic Simulator (SPPARKS) [@Mitchell2023May] offers solely on-lattice kMC modeling capabilities, since the only currently available off-lattice solver is a Metropolis Monte Carlo relaxation scheme. Furthermore, a wide range of on-lattice kMC packages such as kmcos [@kmcos], PyCD [@PyCD], VIS-A-VIS [@VISAVIS], MulSKIPS [@MulSKIPS], Kimocs [@Kimocs], KSOME [@KSOME], kMCpy [@kMCpy], and Morphokinetics [@Morphokinetics] are available in open-source software repositories.
