### 2024-05-20 sntioudis/papreca#12 @liamhuber
Examples: bad path(s?)
A (very simple but) blocking issue for https://github.com/openjournals/joss-reviews/issues/6714

The `papreca_dir="../../build"` variable [in the Simple Adsorption demo](https://github.com/sntioudis/papreca/blob/e0a5929517646a383ff5ed1acc9c6f1dc927f89b/Examples/Simple%20Adsorption/test_simple_adsdep.sh#L3) causes the script to fail by not being able to find the Papreca EXE. I believe this is because of the `cd`, and changing it to `papreca_dir="../../../build"` worked fine for me.

Similarly, I think the double `/` [in the phosphate example](https://github.com/sntioudis/papreca/blob/e0a5929517646a383ff5ed1acc9c6f1dc927f89b/Examples/Phosphate%20Film%20Growth%20from%20TCP%20on%20Fe110/test_phosphates.sh#L2) is going to bork things.
