### 2024-05-20 sntioudis/papreca#13 @liamhuber
Paper: scaling needs data reference
A (very simple but) blocking issue for https://github.com/openjournals/joss-reviews/issues/6714

The paper should indicate (with a hyperlink) that the raw data for the scaling tests is available  [right on the repo](https://github.com/sntioudis/papreca/tree/main/Examples/Phosphate%20Film%20Growth%20from%20TCP%20on%20Fe110/scalability_data). It's kind of tucked away and hidden otherwise.
