### 2024-05-21 sntioudis/papreca#12 @sntioudis
@liamhuber

Since the PAPRECA build folder can be located in different paths for different users, the 'papreca_dir' variable has to be modified to point to the directory of your PAPRECA executable. This is necessary to run any of the examples (as explained in the relevant sections of the documentation). For instance, this is mentioned in the ["Execution section" of the Brownian diffusion example](https://sntioudis.github.io/papreca/md_examples.html):

> Note that, you will have to modify the "papreca_dir" variable in your "test_brownian.sh" script so it points to the path of your PAPRECA executable.

A similar comment was added to the "Execution section" of the Simple adsorption example.

In any case, the "default" 'papreca_dir' value was set to "../../build" because if the user follows the [installation instructions in the documentation](https://sntioudis.github.io/papreca/md_installation.html) without changing the paths or the folder names, then running 'test_brownian.sh' from within the [Brownian Self Diffusion folder](https://github.com/sntioudis/papreca/tree/main/Examples/Brownian%20Self%20Diffusion) will not fail (since the build folder will be located two subdirectories "up"). If the installation procedure is not followed exactly (e.g., the build folder is not created in the cloned PAPRECA repository folder), then the 'papreca_dir' variable will require modification.

> Similarly, I think the double / [in the phosphate example](https://github.com/sntioudis/papreca/blob/e0a5929517646a383ff5ed1acc9c6f1dc927f89b/Examples/Phosphate%20Film%20Growth%20from%20TCP%20on%20Fe110/test_phosphates.sh#L2) is going to bork things.

The double slash is not necessary but will not lead to any issues since consecutive slashes are treated as a single slash in linux bash scripts as stated in the  [Single Unix specification (version 4)](https://pubs.opengroup.org/onlinepubs/9699919799), [base definitions §3.271 pathname](https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_271): “Multiple successive slashes are considered to be the same as one slash, except for the case of exactly two leading <slash> characters.”
