### 2024-05-21 0d2e697
Update paper.bib to include the papreca gitHub page
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
