+++ paper/paper.bib
@@ -1,3 +1,12 @@
+@misc{PAPRECAgitHub,
+  author = {Ntioudis, Stavros and Ewen, James P. and Dini, Daniele and Turner, C. Heath},
+  title = {papreca},
+  year = {2024},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/sntioudis/papreca}
+}
+
 @article{Ntioudis2023Oct,
 	author = {Ntioudis, Stavros and Ewen, James P. and Dini, Daniele and Turner, C. Heath},
 	title = {{A hybrid off-lattice kinetic Monte Carlo/molecular dynamics method for amorphous thin film growth}},
