### 2024-05-21 e3901d7
Update paper.md to include scalability test data link and reference
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
