### 2024-05-21 sntioudis/papreca#13 @sntioudis
A scalability section and a link to the relevant folder containing the test data were added to the paper:

https://github.com/openjournals/joss-reviews/issues/6714#issuecomment-2122298576
