### 2024-05-21 sntioudis/papreca#12 @liamhuber
Indeed, the double / was a-ok.

I'm not talking about brown Ian but rather simple adsorption where the bash script has a loop and a cd. Here the default path is not compliant with the default build directory. The Brownian one is fine.
