### 2024-05-21 1002c7d
Update test_simple_adsdep.sh to prevent conflicts for relative paths
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
