### 2024-05-21 493fbd7
Update run_ALLtests.yml
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
