### 2024-05-21 aba53a8
Update run_ALLtests.yml
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
