+++ .github/workflows/run_ALLtests.yml
@@ -94,6 +94,8 @@ jobs:
           
     - name: Install dependencies
       run: |
+        cd lammpscache/build/
+        ls
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
