+++ .github/workflows/run_ALLtests.yml
@@ -44,7 +44,9 @@ jobs:
       id: cache-lammps-repo
       uses: actions/cache@v4.0.2
       with:
-        path: lammpscache/build/liblammps.a
+        path: |
+          lammpscache/build/liblammps.a
+          lammpscache/build/lmp
         #Key generated from os, and commit HASH. if any of these changes the key will be invalid and a rebuild will be triggered
         key: ${{ runner.os }}-lammps-repo-${{ env.COMMIT_HASH }}
         restore-keys: |
@@ -57,6 +59,9 @@ jobs:
         mkdir build; cd build
         cmake -DPKG_MOLECULE=on -DPKG_RIGID=on -D PKG_QEQ=on -DBUILD_LIB=on -DBUILD_SHARED_LIBS=off -DBUILD_STATIC_LIBS=on ../cmake #Configure LAMMPS, build all required packages to run tests and enable static library building
         cmake --build .
+
+    - name: Verify LAMMPS
+      run: ./lammpscache/build/lmp -h
 
   #Build PAPRECA and PAPRECA test executables on top of LAMMPS.
   #Upload artefacts of both to be used in subsequent tests
@@ -82,7 +87,9 @@ jobs:
       id: cache-lammps
       uses: actions/cache@v4.0.2
       with:
-        path: lammpscache/build/liblammps.a
+        path: |
+          lammpscache/build/liblammps.a
+          lammpscache/build/lmp
         key: ${{ runner.os }}-lammps-repo-${{ env.COMMIT_HASH }}
         restore-keys: |
           ${{ runner.os }}-lammps-repo-
