+++ .github/workflows/run_ALLtests.yml
@@ -138,11 +138,12 @@ jobs:
 
     steps:
 
-    - name: Download all artifacts
-      uses: actions/donwload-artifact@v4
+    - name: Download a single artifact
+      uses: actions/download-artifact@v4
       with:
-        name: '**'
+        name: papreca-artifacts
         path: artifacts
+        merge-multiple: false
 
     - name: check artifacts folder
       run: |
