+++ .github/workflows/run_ALLtests.yml
@@ -121,14 +121,18 @@ jobs:
         mkdir build; cd build
         cmake .. -DLAMMPS_SRC_DIR=$GITHUB_WORKSPACE/mylammps/src/ -DLAMMPS_LIB_DIR=$GITHUB_WORKSPACE/cached/lammps/ -DPAPRECA_SRC_DIR=$GITHUB_WORKSPACE/source
         cmake --build .
+        
+    - name: Move papreca executables to artifacts folder
+      run: |
+        mkdir -p artifacts
+        mv mypapreca/build/papreca artifacts
+        mv $GITHUB_WORKSPACE/tests/source\ tests/build/source_tests artifacts
 
     - name: Upload papreca executables as artifacts
       uses: actions/upload-artifact@v4
       with:
         name: papreca-artifacts
-        path: |
-          mypapreca/build/papreca
-          $GITHUB_WORKSPACE/tests/source\ tests/build/source_tests
+        path: artifacts
 
   testingArtifactDonwload:
   
