### 2024-05-23 89adce7
Update run_tests.sh
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
