+++ tests/source tests/run_tests.sh
@@ -1,1 +1,4 @@
-mpiexec ./build/source_tests -in ./input_files/in_kmc.lmp ./input_files/in_kmc.ppc
+#!/bin/bash
+paprecatests_dir=$1 #Retrieve path from script invocation
+
+mpiexec "${paprecatests_dir}/source_tests" -in ./input_files/in_kmc.lmp ./input_files/in_kmc.ppc
