+++ .github/workflows/run_ALLtests.yml
@@ -144,7 +144,13 @@ jobs:
 
     - name: Checkout #Checkout PAPRECA repository
       uses: actions/checkout@v4 
-      
+
+    - name: Install dependencies
+      run: |
+        sudo apt-get update
+        sudo apt-get install -y build-essential
+        sudo apt-get install mpich
+        
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
@@ -166,8 +172,14 @@ jobs:
     steps:
 
     - name: Checkout #Checkout PAPRECA repository
-      uses: actions/checkout@v4 
+      uses: actions/checkout@v4
       
+    - name: Install dependencies
+      run: |
+        sudo apt-get update
+        sudo apt-get install -y build-essential
+        sudo apt-get install mpich
+        
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
