### 2024-05-23 46b9600
Update test_all_events.sh
Script now aborts if a test fails

Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
