### 2024-05-23 4ced344
Update test_formations.py
Float comparisons are now performed with np.isclose and the script aborts if a test fails

Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
