### 2024-05-23 sntioudis/papreca#11 @sntioudis
All float comparison operations now use numpy.isclose.
