### 2024-05-23 sntioudis/papreca#10 @sntioudis
All the event test python scripts now exit with 1 if the test success rate is 100% and with 0 if the test success rate is less than 100%.

The test_all_events.sh bash script has also been updated to retrieve the relevant python script exit codes. If any of the python scripts returns  1, then the bash script will prematurely exit with code 1 causing the GitHub workflow to fail.
