### 2024-05-23 sntioudis/papreca#7 @sntioudis
The Github workflow file now caches a LAMMPS executable and a LAMMPS library from the latest release branch. The commit hash is used to determine if the cached files require an update. The cached files are used to compile the papreca and source_tests executables.
