### 2024-05-23 sntioudis/papreca#6 @sntioudis
The run_Alltests.yml workflow is now parallelized. A papreca executable is generated and saved as an artifact. Then, all tests run in parallel using the executables saved as artifacts. For the event tests, a matrix is used to concurrently run tests for multiple seeds.
