### 2024-05-23 d736ed3
Update run_ALLtests.yml
Scheduled build and functionality tests to run every day at midnight UTC.

Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
