### 2024-05-23 a109d3b
Update installation.md
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
