### 2024-05-23 sntioudis/papreca#9 @sntioudis
1) The [docs](https://sntioudis.github.io/papreca/md_installation.html) have been updated and mention a specific LAMMPS version (patch_17Apr2024) tested and guaranteed to work with papreca.

2) Build and functionality tests are now scheduled to run every day at midnight UTC. My alerts are enabled for this repository so I will receive an email in case the tests fail.
