### 2024-05-23 sntioudis/papreca#6 @liamhuber
Super nice! I can see on the latest [test report](https://github.com/sntioudis/papreca/actions/runs/9208979405) that these gains are not mind blowing -- I guess you go (approximately) from `sum=5:02` to `slowest=2:14` by parallelizing. About a 20% savings since the bulk of the time (>5:00) is still the Papreca compilation. At any rate, an nice percentage savings and I'm impressed with how quickly you've ramped up the sophistication of your CI.
