### 2024-05-23 sntioudis/papreca#7 @liamhuber
Beautiful, [35s to restore](https://github.com/sntioudis/papreca/actions/runs/9208979405) vs [10min to build from scratch](https://github.com/sntioudis/papreca/actions/runs/9192096995)!
