### 2024-05-23 #6714 @liamhuber
### Update from my end

Everything is looking super good so far. @sntioudis thanks for the quick responses to all my issues and for jumping into the software engineering/devops side of things with both feet! Even most of my non-blocking issues have been resolved and I appreciate the time and effort you've put in to working with GitHub's workflow platform to accomplish this.

The last thing I want to do is try something "new" with the software. I'll also use whatever little toy run I come up with to try and verify the scaling behaviour; I only have access to my desktop machine (8 cores) -- I verified that the scaling example runs OK, but on 4 cores it took overnight to execute so I don't want to repeat that a bunch of times 😂 I'll make sure I wrap that up by the end of next week 👍

### Disclaimers to my review
- The duration of the  project (as a public repo) under the JOSS metrics for "Substantial scholarly effort" is a yellow flag, but since the rest of the metrics seem well met I'm happy to tick this box (e.g. the project was not yet open but is already discussed in a published scientific article by the authors)
- I'm not active in kMC research, so I can't guarantee that no significant citations have been missed, but it's nonetheless clear in the writeup that a solid effort was made to cover the applicable literature; given the good quality of everything else going on in this project, I feel the balance of probability is that they've hit the critical references.
