### 2024-05-23 bb8827b
Update test_all_events.sh
Fixed inconsistent indentation

Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
