### 2024-05-24 962d904
Create build-test-reusable.yml
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
