+++ .github/workflows/templates/build-test-reusable.yml
@@ -1,0 +1,174 @@
+#Template (reusable) workflow that receives a LAMMPS version tag or a LAMMPS branch name.
+#Then, the workflow builds LAMMPS. Finally, the workflow builds PAPRECA and the source_tests.
+#This reusable workflow is utilized for the scheduled build and functionality tests.
+#The workflow can also be used to test for backwards compatibility with different LAMMPS versions.
+
+name: Build and test reusable
+
+on:
+  workflow_call:
+    inputs:
+      clone_type:
+        description: 'Set this to either "branch" or "tag" to clone LAMMPS using a branch or a tag name, respectively'
+        required: true
+        type: string
+      clone_name:
+        description: 'This is the name of either the branch or the tag name'
+        required: true
+        type: string
+      cache_lammps:
+        description: 'Boolean variable. If true the generated LAMMPS library and executable will be cached. If not, they will be saved as artifacts.'
+        required: true
+        type: boolean
+
+jobs:
+
+  buildLAMMPS:
+    runs-on: ubuntu-latest
+
+    steps:
+
+    - name: Clone a specific LAMMPS branch
+      if: ${{ inputs.clone_type == 'branch' }}
+      run: |
+        git clone -b ${{ inputs.clone_name }} https://github.com/lammps/lammps.git mylammps
+        cd mylammps
+        git checkout ${{ inputs.clone_name }}
+        git pull
+
+    - name: Clone a specific LAMMPS tag
+      if: ${{ inputs.clone_type == 'tag' }}
+      run: |
+        git clone --depth 1 --branch ${{ inputs.clone_name }} https://github.com/lammps/lammps.git mylammps
+        
+    - name: Create lammps_artifacts directory to save liblammps.a and lmp
+      run: |
+        mkdir -p lammps_artifacts
+
+    - name: Get LAMMPS release commit SHA
+      if: ${{ inputs.cache_lammps }}
+      id: get-lammps-commit
+      run: |
+        cd mylammps
+        LAMMPS_COMMIT=$(git rev-parse HEAD)
+        echo "COMMIT_HASH=$(git rev-parse HEAD)" >> $GITHUB_ENV
+       
+    - name: Cache LAMMPS library
+      if: ${{ inputs.cache_lammps }}
+      id: cache-lammps-repo
+      uses: actions/cache@v4.0.2
+      with:
+        path: lammps_artifacts
+        #Key generated from os, and commit HASH. if any of these changes the key will be invalid and a rebuild will be triggered
+        key: ${{ runner.os }}-lammps-repo-${{ env.COMMIT_HASH }}
+        restore-keys: |
+          ${{ runner.os }}-lammps-repo-
+
+    - name: Install LAMMPS dependencies
+      if: ( ${{ inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
+      run: |
+        sudo apt-get update
+        sudo apt-get install -y build-essential
+        sudo apt-get install mpich
+        
+    - name: Configure and build LAMMPS
+      if: ( ${{ inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
+      run: |
+        cd mylammps
+        mkdir build; cd build
[…]
