### 2024-05-24 db91d04
Automated tests workflow now build papreca from a reusable workflow
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
