+++ .github/workflows/run_ALLtests.yml
@@ -1,14 +1,6 @@
 name: Build and functionality tests
 
-# 1) Caches LAMMPS's release repository
-# 2) Builds LAMMPS (if changes are detected in the release branch)
-# 3) Builds PAPRECA on top of LAMMPS
-# 4) Builds PAPRECA source tests
-# 5) Saves artifacts of papreca executables
-# After completing steps 1-5, we can be certain that the PAPRECA version can be built without any issues
-# 6) Performs source tests
-# 7) Performs a hybrid kMC/MD test (TCP decomposition on Fe-110)
-# 8) Performs all event tests with 5 different seeds (matrix strategy)
+# Builds LAMMPS from a reusable workflow and then executes all PAPRECA tests
 
 
 on:
@@ -16,129 +8,17 @@ on:
     - cron: "0 0 * * *"
 
 jobs:
-
-  buildLAMMPS:
-    runs-on: ubuntu-latest
-
-    steps:
-    
-    - name: Clone LAMMPS repository (release branch)
-      run: |
-        git clone -b release https://github.com/lammps/lammps.git mylammps
-        cd mylammps
-        git checkout release
-        git pull
-
-    - name: Get LAMMPS release commit SHA
-      id: get-lammps-commit
-      run: |
-        cd mylammps
-        LAMMPS_COMMIT=$(git rev-parse HEAD)
-        echo "COMMIT_HASH=$(git rev-parse HEAD)" >> $GITHUB_ENV
-
-    - name: Create cached directory to save liblammps.a and lmp
-      run: |
-        mkdir -p cached
-        cd ./cached
-        mkdir -p lammps
-        
-    - name: Cache LAMMPS library
-      id: cache-lammps-repo
-      uses: actions/cache@v4.0.2
-      with:
-        path: cached/lammps
-        #Key generated from os, and commit HASH. if any of these changes the key will be invalid and a rebuild will be triggered
-        key: ${{ runner.os }}-lammps-repo-${{ env.COMMIT_HASH }}
-        restore-keys: |
-          ${{ runner.os }}-lammps-repo-
-
-    - name: Install LAMMPS dependencies
-      if: steps.cache-lammps-repo.outputs.cache-hit != 'true'
-      run: |
-        sudo apt-get update
-        sudo apt-get install -y build-essential
-        sudo apt-get install mpich
-        
-    - name: Configure and build LAMMPS
-      if: steps.cache-lammps-repo.outputs.cache-hit != 'true'
-      run: |
-        cd mylammps
-        mkdir build; cd build
-        cmake -DPKG_MOLECULE=on -DPKG_RIGID=on -D PKG_QEQ=on -DBUILD_LIB=on -DBUILD_SHARED_LIBS=off -DBUILD_STATIC_LIBS=on ../cmake #Configure LAMMPS, build all required packages to run tests and enable static library building
-        cmake --build .
-
-    - name: Verify LAMMPS
-      if: steps.cache-lammps-repo.outputs.cache-hit != 'true'
-      run: ./mylammps/build/lmp -h
-
-    - name: Move lmp executable and lammps library to cached directory
-      if: steps.cache-lammps-repo.outputs.cache-hit != 'true'
-      run: |
-        mv mylammps/build/lmp cached/lammps
-        mv mylammps/build/liblammps.a cached/lammps
-
-  buildPAPRECA: 
-
-    needs: buildLAMMPS
-    
-    runs-on: ubuntu-latest
-
-    steps:
-
-    - name: Checkout
-      uses: actions/checkout@v4 #Checkout PAPRECA repository
-
[…]
