### 2024-05-24 bdb2181
Enabled manual trigger for automated tests (on top of scheduled triggers)
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
