+++ .github/workflows/run_ALLtests.yml
@@ -6,6 +6,7 @@ name: Build and functionality tests
 on:
   schedule: #Test runs everyday at midnight UTC
     - cron: "0 0 * * *"
+  workflow_dispatch:
 
 jobs:
   build-reusable:
