+++ .github/workflows/build-reusable.yml
@@ -65,14 +65,14 @@ jobs:
           ${{ runner.os }}-lammps-repo-
 
     - name: Install LAMMPS dependencies
-      if: ( ${{ inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
+      if: ${{ ( inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
       run: |
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
         
     - name: Configure and build LAMMPS
-      if: ( ${{ inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
+      if: ${{ ( inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
       run: |
         cd mylammps
         mkdir build; cd build
@@ -80,11 +80,11 @@ jobs:
         cmake --build .
 
     - name: Verify LAMMPS
-      if: ( ${{ inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
+      if: ${{ ( inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
       run: ./mylammps/build/lmp -h
 
     - name: Move lmp executable and lammps library to lammps_artifacts directory
-      if: ( ${{ inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
+      if: ${{ ( inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
       run: |
         mv mylammps/build/lmp lammps_artifacts
         mv mylammps/build/liblammps.a lammps_artifacts
