### 2024-05-24 3945955
Create lammps-compatibility.yml
Testing backwards compatibility of PAPRECA with various LAMMPS versions using a matrix strategy.

Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
