### 2024-05-24 c7dde66
Update lammps-compatibility.yml
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
