### 2024-05-24 9bcefea
Included commit hashes to artifact folder names to prevent conflicts
Conflicts were happening in lammps-compatibility tests when the action was trying to save multiple different LAMMPS builds at the same directory.

Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
