+++ .github/workflows/build-reusable.yml
@@ -49,6 +49,9 @@ jobs:
         LAMMPS_COMMIT=$(git rev-parse HEAD)
         echo "COMMIT_HASH=$(git rev-parse HEAD)" >> $GITHUB_ENV
         
+    - name: Sanity check
+      run: echo "${{ env.COMMIT_HAS }}"
+      
     - name: Create lammps_artifacts directory to save liblammps.a and lmp
       run: |
         mkdir -p lammps_artifacts_${{ env.COMMIT_HASH }}
