+++ .github/workflows/build-reusable.yml
@@ -49,21 +49,20 @@ jobs:
       id: get-lammps-commit
       run: |
         cd mylammps
-        LAMMPS_HASH=$(git rev-parse HEAD)
-        echo "COMMIT_HASH=${LAMMPS_HASH}" >> $GITHUB_ENV
+        echo "COMMIT_HASH=$(git rev-parse HEAD)" >> $GITHUB_ENV
       
     - name: Create lammps_artifacts directory to save liblammps.a and lmp
       run: |
-        mkdir -p lammps_artifacts_${{ LAMMPS_HASH }}
+        mkdir -p lammps_artifacts_${{ inputs.clone_name }}
        
     - name: Cache LAMMPS library
       if: ${{ inputs.cache_lammps }}
       id: cache-lammps-repo
       uses: actions/cache@v4.0.2
       with:
-        path: lammps_artifacts_${{ LAMMPS_HASH }}
+        path: lammps_artifacts_${{ inputs.clone_name }}
         #Key generated from os, and commit HASH. if any of these changes the key will be invalid and a rebuild will be triggered
-        key: ${{ runner.os }}-lammps-repo-${{ LAMMPS_HASH }}
+        key: ${{ runner.os }}-lammps-repo-${{ env.COMMIT_HASH }}
         restore-keys: |
           ${{ runner.os }}-lammps-repo-
 
@@ -89,15 +88,15 @@ jobs:
     - name: Move lmp executable and lammps library to lammps_artifacts directory
       if: ${{ ( inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
       run: |
-        mv mylammps/build/lmp lammps_artifacts_${{ LAMMPS_HASH }}
-        mv mylammps/build/liblammps.a lammps_artifacts_${{ LAMMPS_HASH }}
+        mv mylammps/build/lmp lammps_artifacts_${{ inputs.clone_name }}
+        mv mylammps/build/liblammps.a lammps_artifacts_${{ inputs.clone_name }}
 
     - name: Save LAMMPS as an artifact (if chosen by user)
       if: ${{ !inputs.cache_lammps }} #This means that we do not use a LAMMPS cache so we save it as an artifact
       uses: actions/upload-artifact@v4
       with:
-        name: lammps-artifacts_${{ LAMMPS_HASH }}
-        path: lammps_artifacts_${{ LAMMPS_HASH }}
+        name: lammps-artifacts_${{ inputs.clone_name }}
+        path: lammps_artifacts_${{ inputs.clone_name }}
       
   buildPAPRECA: 
 
@@ -128,7 +127,7 @@ jobs:
       id: cache-lammps
       uses: actions/cache@v4.0.2
       with:
-        path: lammps_artifacts_${{ env.COMMIT_HASH }}
+        path: lammps_artifacts_${{ inputs.clone_name }}
         #Key generated from os, and commit HASH. if any of these changes the key will be invalid and a rebuild will be triggered
         key: ${{ runner.os }}-lammps-repo-${{ env.COMMIT_HASH }}
         restore-keys: |
@@ -138,8 +137,8 @@ jobs:
       if: ${{ !inputs.cache_lammps }}
       uses: actions/download-artifact@v4
       with:
-        name: lammps-artifacts_${{ env.COMMIT_HASH }}
-        path: lammps_artifacts_${{ env.COMMIT_HASH }}
+        name: lammps-artifacts_${{ inputs.clone_name }}
[…]
