+++ .github/workflows/run_ALLtests.yml
@@ -37,15 +37,15 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.COMMIT_HASH }} #use commit hash (enviromental variable captured in the reusable workflow)
-        path: papreca_artifacts_${{ env.COMMIT_HASH }}
+        name: papreca-artifacts_${{ env.clone_name }} #use commit hash (enviromental variable captured in the reusable workflow)
+        path: papreca_artifacts_${{ env.clone_name }}
         merge-multiple: false
 
     - name: Run source tests
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.COMMIT_HASH }}/source_tests
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/source_tests
         cd $GITHUB_WORKSPACE/tests/source\ tests/
-        bash run_tests.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.COMMIT_HASH }}
+        bash run_tests.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}
 
   hybridTest:
   
@@ -67,16 +67,16 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.COMMIT_HASH }}
-        path: papreca_artifacts_${{ env.COMMIT_HASH }}
+        name: papreca-artifacts_${{ env.clone_name }}
+        path: papreca_artifacts_${{ env.clone_name }}
         merge-multiple: false
         
 
     - name: Run hybrid test
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.COMMIT_HASH }}/papreca
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/papreca
         cd $GITHUB_WORKSPACE/tests/hybrid\ kMC\ MD/
-        bash ./test_hybrid.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.COMMIT_HASH }}
+        bash ./test_hybrid.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}
 
 
   eventTests:
@@ -110,14 +110,14 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.COMMIT_HASH }}
-        path: papreca_artifacts_${{ env.COMMIT_HASH }}
+        name: papreca-artifacts_${{ env.clone_name }}
+        path: papreca_artifacts_${{ env.clone_name }}
         merge-multiple: false
         
 
     - name: Run event test
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.COMMIT_HASH }}/papreca
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/papreca
         echo -e "Starting all event tests with seed: ${{ matrix.seed }}"
         cd $GITHUB_WORKSPACE/tests/event\ tests/
-        bash test_all_events.sh ${{ matrix.seed }} $GITHUB_WORKSPACE/papreca_artifacts_${{ env.COMMIT_HASH }} python3
+        bash test_all_events.sh ${{ matrix.seed }} $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }} python3
