+++ .github/workflows/build-reusable.yml
@@ -101,7 +101,7 @@ jobs:
 
     - name: Export clone name as env variable
       run: |
-        echo "clone_name=${inputs.clone_name}" >> $GITHUB_ENV #In order to use it in subsequent jobs to retrieve paths
+        echo "clone_name=${{inputs.clone_name}}" >> $GITHUB_ENV #In order to use it in subsequent jobs to retrieve paths
       
   buildPAPRECA: 
 
