+++ .github/workflows/run_ALLtests.yml
@@ -15,11 +15,21 @@ jobs:
      clone_type: branch
      clone_name: release
      cache_lammps: true
-   
+
+  retrieveCloneName:
+
+    needs: build-reusable
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Retrieve clone_name
+        run: |
+          echo "clone_name=${{ needs.build-reusable.outputs.clone_name }}" >> $GITHUB_ENV
 
   sourceTests:
   
-    needs: build-reusable
+    needs: retrieveCloneName
     
     runs-on: ubuntu-latest
 
@@ -49,7 +59,7 @@ jobs:
 
   hybridTest:
   
-    needs: build-reusable
+    needs: retrieveCloneName
     
     runs-on: ubuntu-latest
 
@@ -81,7 +91,7 @@ jobs:
 
   eventTests:
   
-    needs: build-reusable
+    needs: retrieveCloneName
     
     runs-on: ubuntu-latest
 
