+++ .github/workflows/build-reusable.yml
@@ -184,4 +184,4 @@ jobs:
     steps:
       - name: Export clone name
         id: export-clone-name
-        run: echo "clone_name=${{ inputs.clone_name }}" >> $GITHUB_OUTPUTS
+        run: echo "clone_name=${ inputs.clone_name }" >> $GITHUB_OUTPUTS
