+++ .github/workflows/build-reusable.yml
@@ -171,17 +171,3 @@ jobs:
       with:
         name: papreca-artifacts_${{ inputs.clone_name }}
         path: papreca_artifacts_${{ inputs.clone_name }}
-
-  exportCloneName: 
-
-    needs: buildLAMMPS
-
-    runs-on: ubuntu-latest
-
-    outputs:
-      CLONE_NAME: ${{ steps.export-clone-name.outputs.CLONE_NAME }}
-
-    steps:
-      - name: Export clone name
-        id: export-clone-name
-        run: echo "CLONE_NAME=${{ inputs.clone_name }}" >> $GITHUB_OUTPUTS
