+++ .github/workflows/run_ALLtests.yml
@@ -9,27 +9,26 @@ on:
   workflow_dispatch:
 
 jobs:
+
+  setCloneName:
+  
+    runs-on: ubuntu-latest
+    
+    steps:
+      - name: Set github repo target name
+        run: echo "CLONE_NAME=release" >> $GITHUB_ENV
+    
   build-reusable:
+   needs: setCloneName
    uses: sntioudis/papreca/.github/workflows/build-reusable.yml@main
    with:
      clone_type: branch
-     clone_name: release
+     clone_name: ${{ env.CLONE_NAME }}
      cache_lammps: true
-
-  retrieveCloneName:
-
-    needs: build-reusable
-
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Retrieve clone_name
-        run: |
-          echo "clone_name=${{ needs.build-reusable.outputs.clone_name }}" >> $GITHUB_ENV
 
   sourceTests:
   
-    needs: retrieveCloneName
+    needs: build-reusable
     
     runs-on: ubuntu-latest
 
@@ -47,19 +46,19 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.clone_name }} #use commit hash (enviromental variable captured in the reusable workflow)
-        path: papreca_artifacts_${{ env.clone_name }}
+        name: papreca-artifacts_${{ env.CLONE_NAME }} #use commit hash (enviromental variable captured in the reusable workflow)
+        path: papreca_artifacts_${{ env.CLONE_NAME }}
         merge-multiple: false
 
     - name: Run source tests
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/source_tests
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.CLONE_NAME }}/source_tests
         cd $GITHUB_WORKSPACE/tests/source\ tests/
-        bash run_tests.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}
+        bash run_tests.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.CLONE_NAME }}
 
   hybridTest:
   
-    needs: retrieveCloneName
+    needs: build-reusable
     
     runs-on: ubuntu-latest
 
@@ -77,21 +76,21 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.clone_name }}
-        path: papreca_artifacts_${{ env.clone_name }}
+        name: papreca-artifacts_${{ env.CLONE_NAME }}
+        path: papreca_artifacts_${{ env.CLONE_NAME }}
         merge-multiple: false
         
 
     - name: Run hybrid test
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/papreca
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.CLONE_NAME }}/papreca
         cd $GITHUB_WORKSPACE/tests/hybrid\ kMC\ MD/
-        bash ./test_hybrid.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}
+        bash ./test_hybrid.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.CLONE_NAME }}
 
 
   eventTests:
   
-    needs: retrieveCloneName
+    needs: build-reusable
     
     runs-on: ubuntu-latest
 
@@ -120,14 +119,14 @@ jobs:
[…]
