+++ .github/workflows/run_ALLtests.yml
@@ -8,22 +8,16 @@ on:
     - cron: "0 0 * * *"
   workflow_dispatch:
 
-jobs:
-
-  setCloneName:
+env:
+  clone_name: release
   
-    runs-on: ubuntu-latest
-    
-    steps:
-      - name: Set github repo target name
-        run: echo "CLONE_NAME=release" >> $GITHUB_ENV
+jobs:
     
   build-reusable:
-   needs: setCloneName
    uses: sntioudis/papreca/.github/workflows/build-reusable.yml@main
    with:
      clone_type: branch
-     clone_name: ${{ env.CLONE_NAME }}
+     clone_name: ${{ env.clone_name }}
      cache_lammps: true
 
   sourceTests:
@@ -46,15 +40,15 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.CLONE_NAME }} #use commit hash (enviromental variable captured in the reusable workflow)
-        path: papreca_artifacts_${{ env.CLONE_NAME }}
+        name: papreca-artifacts_${{ env.clone_name }} #use commit hash (enviromental variable captured in the reusable workflow)
+        path: papreca_artifacts_${{ env.clone_name }}
         merge-multiple: false
 
     - name: Run source tests
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.CLONE_NAME }}/source_tests
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/source_tests
         cd $GITHUB_WORKSPACE/tests/source\ tests/
-        bash run_tests.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.CLONE_NAME }}
+        bash run_tests.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}
 
   hybridTest:
   
@@ -76,16 +70,16 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.CLONE_NAME }}
-        path: papreca_artifacts_${{ env.CLONE_NAME }}
+        name: papreca-artifacts_${{ env.clone_name }}
+        path: papreca_artifacts_${{ env.clone_name }}
         merge-multiple: false
         
 
     - name: Run hybrid test
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.CLONE_NAME }}/papreca
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/papreca
         cd $GITHUB_WORKSPACE/tests/hybrid\ kMC\ MD/
-        bash ./test_hybrid.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.CLONE_NAME }}
+        bash ./test_hybrid.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}
 
 
   eventTests:
@@ -119,14 +113,14 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.CLONE_NAME }}
-        path: papreca_artifacts_${{ env.CLONE_NAME }}
+        name: papreca-artifacts_${{ env.clone_name }}
+        path: papreca_artifacts_${{ env.clone_name }}
         merge-multiple: false
         
 
     - name: Run event test
       run: |
[…]
