+++ .github/workflows/run_ALLtests.yml
@@ -17,7 +17,7 @@ jobs:
    uses: sntioudis/papreca/.github/workflows/build-reusable.yml@main
    with:
      clone_type: branch
-     clone_name: ${{ env.clone_name }}
+     clone_name: ${{ clone_name }}
      cache_lammps: true
 
   sourceTests:
