+++ .github/workflows/run_ALLtests.yml
@@ -49,15 +49,15 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.clone_name }} #use commit hash (enviromental variable captured in the reusable workflow)
-        path: papreca_artifacts_${{ env.clone_name }}
+        name: papreca-artifacts_${{ needs.setCloneName.outputs.clone_name }} #use commit hash (enviromental variable captured in the reusable workflow)
+        path: papreca_artifacts_${{ needs.setCloneName.outputs.clone_name }}
         merge-multiple: false
 
     - name: Run source tests
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/source_tests
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ needs.setCloneName.outputs.clone_name }}/source_tests
         cd $GITHUB_WORKSPACE/tests/source\ tests/
-        bash run_tests.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}
+        bash run_tests.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ needs.setCloneName.outputs.clone_name }}
 
   hybridTest:
   
@@ -79,16 +79,16 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.clone_name }}
-        path: papreca_artifacts_${{ env.clone_name }}
+        name: papreca-artifacts_${{ needs.setCloneName.outputs.clone_name }}
+        path: papreca_artifacts_${{ needs.setCloneName.outputs.clone_name }}
         merge-multiple: false
         
 
     - name: Run hybrid test
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/papreca
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ needs.setCloneName.outputs.clone_name }}/papreca
         cd $GITHUB_WORKSPACE/tests/hybrid\ kMC\ MD/
-        bash ./test_hybrid.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}
+        bash ./test_hybrid.sh $GITHUB_WORKSPACE/papreca_artifacts_${{ needs.setCloneName.outputs.clone_name }}
 
 
   eventTests:
@@ -122,14 +122,14 @@ jobs:
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
       with:
-        name: papreca-artifacts_${{ env.clone_name }}
-        path: papreca_artifacts_${{ env.clone_name }}
+        name: papreca-artifacts_${{ needs.setCloneName.outputs.clone_name }}
+        path: papreca_artifacts_${{ needs.setCloneName.outputs.clone_name }}
         merge-multiple: false
         
 
     - name: Run event test
       run: |
-        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }}/papreca
+        chmod +rwx $GITHUB_WORKSPACE/papreca_artifacts_${{ needs.setCloneName.outputs.clone_name }}/papreca
         echo -e "Starting all event tests with seed: ${{ matrix.seed }}"
         cd $GITHUB_WORKSPACE/tests/event\ tests/
-        bash test_all_events.sh ${{ matrix.seed }} $GITHUB_WORKSPACE/papreca_artifacts_${{ env.clone_name }} python3
+        bash test_all_events.sh ${{ matrix.seed }} $GITHUB_WORKSPACE/papreca_artifacts_${{ needs.setCloneName.outputs.clone_name }} python3
