### 2024-05-24 915481e
Disabled fail-fast for lammps compatibility tests
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
