### 2024-05-24 db833cb
Update lammps-compatibility.yml
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
