+++ .github/workflows/lammps-compatibility.yml
@@ -11,7 +11,7 @@ jobs:
     strategy:
       matrix:
         clone_name: [ patch_17Apr2024 , patch_7Feb2024 , patch_21Nov2023 , patch_2Aug2023 , patch_15Jun2023 , patch_28Mar2023 , patch_8Feb2023 , patch_22Dec2022 , patch_3Nov2022 , patch_15Sep2022 , patch_3Aug2022] #Target many feature releases from newest to oldest
-        fail-fast: false #disable this because if one fails, the whole workflow will fail
+      fail-fast: false #disable this because if one fails, the whole workflow will fail
     uses: sntioudis/papreca/.github/workflows/build-reusable.yml@main
     with:
       clone_type: tag
