### 2024-05-24 bd9f3a4
Update installation.md
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
