### 2024-05-24 6f023cc
Included compatibility badge
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
