### 2024-05-24 sntioudis/papreca#8 @liamhuber
Very nice! 🚀 🚀

I think you can probably streamline the reusable workflow using https://github.com/actions/checkout. I've never used it to check out a repo other than the one I was working in, but it looks like they support that a-ok. Then it looks to me like you can supply the `clone_name` as their `ref`, and their action will handle the `clone_type` and all the git code. You might need to adjust some paths in the rest of the workflow, as I believe what you have as `./mylammps` will just wind up as `.` (actually I just read deeper, and it looks like you could even pass it the `path: mylammps` input argument to keep things the same if you want). This is all just polish though, the functionality you introduced is excellent as-is.
