### 2024-05-29 sntioudis/papreca#4 @sntioudis
It looks like the _LAMMPS static library (liblammps.a)_ is not built if LAMMPS is installed through the [conda-forge route](https://docs.lammps.org/Install_conda.html). I just verified that by building LAMMPS on my workstation through conda-forge.

Papreca uses the _static LAMMPS library_, hence, I reckon that a new lammps-static feedstock would have to be made.
