### 2024-05-31 #6714 @liamhuber
@srmnitc, I'm done and satisfied. @sntioudis, :rocket: nice work!

I made my own very small scaling test and was only able to run on 1,2,4 cores, but it looks qualitatively similar to the paper figure. The key thing is that the expensive atomic calculations leverage all the parallelization of lammps, and that's working a-ok.

My "new" thing was not mindblowing, but I had fun. I wanted to explore Ostwald ripening, so I modified the brownian motion example to include a minimization (full inputs collapsed below). The diffusive hops are a little bit aphysical for this purpose, but I suppose one could very roughly think of it as interacting particles moving around in a non-interacting, implicit, amorphous medium -- i.e. the particles need to wait for the medium to provide diffusion opportunities but don't impact the energetics. I still expect Ostwald ripening under these conditions even just because of the surface-area-to-volume statistics about who's most likely to lose a % of itself to the "gas".

<summary>Papreca input</summary>

```
#PAPRECA input file

#Basic settings
KMC_steps 4000000
KMC_per_MD 100
random_seed 13452234
fluid_atomtypes 1 1

#Sigmas for types
sigmas_options LAMMPS mix geom

#Equilibration
trajectory_duration 1
minimize_prior yes minimize 1.0e-16 1.0e-16 100 20000

#Predefined events
create_DiffusionHop 1 0.0 3 yes 1 rate_manual 1.0
random_diffvecs yes 3D

#NeibLists
neiblists lj/cut zero
```

<summary>Lammps input</summary>

```
#LAMMPS input file

#---------------------------Simulation parameters--------------------------
dimension 3
units           real
atom_style      full
pair_style 	hybrid/overlay lj/cut 10 zero 10 full

#---------------------------Processors-------------------------------------
processors * * *
#---------------------------Define boundaries------------------------------
boundary p p p

#--------------------------Create simulation-------------------------------
region simulation_box block -50 50 -50 50 -50 50
create_box 1 simulation_box

#--------------------------Create atoms------------
create_atoms 1 random 1000 141241 NULL
#---------------------------Atom masses------------------------------------
mass 1    1.0079
#----------------------Pair coeffs -------------------------
pair_coeff 1 1 lj/cut 0.0152 2.84642

pair_coeff * * zero

#----------------------Define group and neib/charge interactions-----------
neighbor        2 bin
neigh_modify    every 1 delay 0 check yes

group fluid type 2
group frozen type 1
#----------------Set Timestep----------------------------------------------
timestep    	1.0e-100
#----------------------File Outputs----------------------------------------
dump 1 all xyz 404 ostwald.xyz
dump_modify 1 element H first yes
```

I was a little surprised to see my particles come out as needles rather than spheroids, but it's probably just something to do with the LJ trying to pack itself amorphously and I'm not stressed about it. Overall the ripening proceeded as expected and I had fun:

Initial:
![start](https://github.com/openjournals/joss-reviews/assets/20283329/7074f2e1-38e6-49ce-93e8-c48b3d945b48)

Early:
![early](https://github.com/openjournals/joss-reviews/assets/20283329/45996df1-7146-404c-b19b-6a36c33af0fb)

Mid:
![mid](https://github.com/openjournals/joss-reviews/assets/20283329/b3fb882e-b82f-41c9-a72b-470b585df023)

Late:
![late](https://github.com/openjournals/joss-reviews/assets/20283329/b02bdb1b-092f-4462-99f3-855e66d74418)

Some concluding thoughts:
- The docs are well structured and sufficiently detailed, but I found the examples great to learn from. As time goes on, keep adding more of those!
- It's nice that the phosphate scaling test is directly available, but when getting started and running through all the examples it came as a nasty shock to go from the example finishing in a few seconds to hours -- maybe break it into a separate folder or scale down the run-time in the default script?
- Extending the docs with atomistic schematic diagrams for the different event types would be cool
- I had no trouble getting the trajectory file to dump how I wanted, but my `log.lammps` files got huge --  is there a way to control this from the lammps or papreca inputs? It's been years since I worked with the lammps input script directly, so it's highly probably that I'm just being a big dummy here, but the world is full of dummies and some hand-holding documentation on controlling output file size (even if it's just regurgitating lammps docs) might be helpful
