### 2024-06-03 #6714 @jfaraudo
@srmnitc Yes, I have to reinstall LAMMPS in my machine since I have a version older than that recommended for the build of PAPRECA (LAMMPS is required as a library). I am unsure to do the the minimal LAMMPS installation required for PAPRECA or my usual build of LAMMPS which adds more features not needed in PAPRECA. Perhaps I'll consider using a fresh spare computer for testing. Sorry again for that delay.
