### 2024-06-03 #6714 @sntioudis
@liamhuber really thank you for your time and your commitment to this review! Thanks to your suggestions, the software and the paper have significantly improved over the past few weeks. I also had good fun working on the repository and I learned a lot! Furthermore, thank you for using papreca to study something different than the documented examples. It was a joy to see the software producing sensible results for a system that has never been studied before with papreca.

> It's nice that the phosphate scaling test is directly available, but when getting started and running through all the examples it came as a nasty shock to go from the example finishing in a few seconds to hours -- maybe break it into a separate folder or scale down the run-time in the default script?

This is a fair point. I will add a disclaimer to the documentation page to warn users that the phosphate examples require considerably more time to finish than the rest of the examples. I will also test if it is possible to reduce the trajectory duration and still obtain sensible growth results.

> Extending the docs with atomistic schematic diagrams for the different event types would be cool.

Another good point. I will add a few images to the documentation (probably in the "Essential Information" or the "PAPRECA Commands" section).

> I had no trouble getting the trajectory file to dump how I wanted, but my log.lammps files got huge -- is there a way to control this from the lammps or papreca inputs? It's been years since I worked with the lammps input script directly, so it's highly probably that I'm just being a big dummy here, but the world is full of dummies and some hand-holding documentation on controlling output file size (even if it's just regurgitating lammps docs) might be helpful

Adding the following line to your LAMMPS input prevents LAMMPS from constantly writing on the log.lammps file:

`log none'

I tried adding this line to the end of your LAMMPS input (Oswald ripening example) and the size of the output log.lammps file was reduced to just 1.7 kb.

I will soon add a note in the documentation to encourage users to include the 'log none' option in their papreca simulations to prevent log.lammps files from getting very large!
