### 2024-06-03 02f376c
Added suggestion to use log.lammps in LAMMPS input files.
This is to prevent log.lammps files from getting very large.

Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
