### 2024-06-03 fe59acc
Added disclaimer for the total runtime of the phosphates example.
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
