### 2024-06-03 1bd4af6
Updated to peaceiris/actions-gh-pages@v4
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
