### 2024-06-03 bc9e29f
Included event schematics for essentials.md
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
