### 2024-06-03 d425ee4
Fixed event schematic paths for documentation
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
