+++ tests/README.md
@@ -1,7 +1,7 @@
 PAPRECA functionality tests
 ========
 
-Tests consist of two main families:
+Tests consist of three main families:
 
 - Event tests (located in ./event\ tests/ directory)
 - Source tests (located in ./source\ tests/)
