### 2024-06-03 #6714 @jfaraudo
I did a fresh install of LAMMPS with cmake and it is now up to date . But I am still having problems installing PAPRECA with cmake in my desktop ubuntu workstation, apparently a conflict with anaconda libraries. I think I'll try with another computer and see what happens.
This is the error, just in case the solution is obvious to anybody

-- Configuring done (0.6s)
CMake Warning at CMakeLists.txt:51 (add_executable):
  Cannot generate a safe runtime search path for target papreca because files
  in some directories may conflict with libraries in implicit directories:

    runtime library [libpng16.so.16] in /usr/lib/x86_64-linux-gnu may be hidden by files in:
      /home/jfaraudo/anaconda3/lib
    runtime library [libz.so.1] in /usr/lib/x86_64-linux-gnu may be hidden by files in:
      /home/jfaraudo/anaconda3/lib
    runtime library [libgomp.so.1] in /usr/lib/gcc/x86_64-linux-gnu/7 may be hidden by files in:
      /home/jfaraudo/anaconda3/lib

  Some of these libraries may not be found correctly.

-- Generating done (0.0s)
-- Build files have been written to: /home/jfaraudo/source_programs/papreca-main/build
