### 2024-06-03 #6714 @sntioudis
@jfaraudo I am not 100% sure but it appears that there are duplicate installations of certain shared libraries in your system. This issue is discussed [here](https://stackoverflow.com/questions/22536567/cmake-cannot-resolve-runtime-directory-path). This looks like a warning (rather than an error), which means that building the project might work (however, the produced executable might lead to unpredictable behavior - so it is better to fix that before testing papreca). Also, I noticed that there are conflicts between the 'user/lib/' and 'anaconda3/lib' paths which might be a conda issue (e.g., some of those libraries were installed on the conda 'base' environment or there were conflicts on the conda environment you used to install papreca? Not sure about that, just an idea).

In the provided link above, one of the suggestions was to uninstall one of the two library installations (located in the 2 different locations indicated by your error). Another solution might be to provide cmake with a hint, so it targets a specific library location.

Of course, I understand that tackling the warning and uninstalling libraries might be harder than installing LAMMPS and PAPRECA on another computer. In any case, keep me updated and I will be very happy to assist you further if neccessary!
