### 2024-06-04 sntioudis/papreca#15 @jfaraudo
Comments to paper
The paper looks nice. I have only a few comments:
- This code implements hybrid kMC/MD simulations by combining the new kMC code with the LAMMPS MD engine. The concepts reminds me a bit that of py-MCMD (code : https://github.com/GOMC-WSU/py-MCMD paper: J. Chem. Theory Comput. 2022, 18, 8, 4983–4994 , https://pubs.acs.org/doi/10.1021/acs.jctc.1c00911 ) . In this previous code the authors combine MC with MD using the NAMD engine (instead of LAMMPS). An important difference is that in py_MCMD the MC method is a Grand Canonical MC , aimed at adding/removing molecules and atoms from the system following Grand Canonical ensemble. Here the method is kMC , aimed at the speed up of kinetic processes that are too slow in a direct MD simulation. I think this previous method has to be mentioned (and very briefly compared with the present approach) in the paper.
- The authors mention at several points the generality of the method, and in the statement of need they include examples such as protein folding or enzyme reactions that require consideration of solvent, in order to be simulated properly. I am not completelly sure that the present method works efficiently in presence of solvent. In my experience, explicit solvent makes MC methods quite inefficient (pls note that I have still to run the examples, so I admit that I can be wrong). For me it is clear that in situations such as solid state simulations , molecules in contact with solids and thinks like this the method is perfect but I think that a comment whether this method is recommended or not in simulations with explicit solvent is in order.
