### 2024-06-04 #6714 @jfaraudo
I have added a few comments about the paper in this issue: https://github.com/sntioudis/papreca/issues/15
