### 2024-06-04 38f1f21
Added FFTW to dependencies
Some LAMMPS packages (e.g., KSPACE) require FFTW. Building the PAPRECA (on top of LAMMPS) without FFTW can lead to linkage errors.

Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
