+++ .github/workflows/build-reusable.yml
@@ -69,6 +69,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
+        sudo apt-get install fftw3
         
     - name: Configure and build LAMMPS
       if: ${{ ( inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
@@ -143,6 +144,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
+        sudo apt-get install fftw3
 
     - name: Build PAPRECA
       run: |
