+++ .github/workflows/run_ALLtests.yml
@@ -47,6 +47,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
+        sudo apt-get install fftw3
         
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
@@ -79,6 +80,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
+        sudo apt-get install fftw3
         
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
@@ -124,6 +126,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
+        sudo apt-get install fftw3
         
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
