+++ .github/workflows/build-reusable.yml
@@ -69,7 +69,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
-        sudo apt-get install fftw3
+        sudo apt-get install libfftw3-dev
         
     - name: Configure and build LAMMPS
       if: ${{ ( inputs.cache_lammps && steps.cache-lammps-repo.outputs.cache-hit != 'true' ) || !inputs.cache_lammps }}
@@ -144,7 +144,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
-        sudo apt-get install fftw3
+        sudo apt-get install libfftw3-dev
 
     - name: Build PAPRECA
       run: |
