+++ .github/workflows/run_ALLtests.yml
@@ -47,7 +47,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
-        sudo apt-get install fftw3
+        sudo apt-get install libfftw3-dev
         
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
@@ -80,7 +80,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
-        sudo apt-get install fftw3
+        sudo apt-get install libfftw3-dev
         
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
@@ -126,7 +126,7 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
-        sudo apt-get install fftw3
+        sudo apt-get install libfftw3-dev
         
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
