### 2024-06-04 43d6b7e
Added KSPACE package to test FFTW lib
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
