+++ .github/workflows/run_ALLtests.yml
@@ -47,7 +47,6 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
-        sudo apt-get install libfftw3-dev
         
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
@@ -80,7 +79,6 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
-        sudo apt-get install libfftw3-dev
         
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
@@ -126,7 +124,6 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential
         sudo apt-get install mpich
-        sudo apt-get install libfftw3-dev
         
     - name: Restore papreca executables
       uses: actions/download-artifact@v4
