### 2024-06-05 sntioudis/papreca#16 @jfaraudo
Installation comments: Add note for mpicxx compiler in documentation (i.e., define DCMAKE_CXX_COMPILER for cmake)
I tried to install papreca in a machnie with a fresh install of Fedora Linux.
LAMMPS library was build as a fresh install following papreca documentation instructions , this is OK.

Now, when trying ti build papreca with cmak I had the following problem:

First, I did the cmake without errors or warnings:

cmake ../Installation/CMake -DLAMMPS_SRC_DIR=/home/jfaraudo/src -DLAMMPS_LIB_DIR=/home/jfaraudo/mylammps/build
-- Found LAMMPS library: /home/jfaraudo/mylammps/build/liblammps.a
-- Configuring additional packages...
CMake Warning at CMakeLists.txt:59 (message):
  PAPRECA does not use OpenMP, JPEG, and PNG packages.  However, the LAMMPS
  library uses them and therefore, we need to compile against those libraries
  to avoid linkage errors.  Depending on how YOUR LAMMPS library was built
  (i.e., based on the LAMMPS packages you enabled) you might need to enable
  additional libraries in the PAPRECA CMakeLists.txt file.

-- Configuring done (1.3s)
-- Generating done (0.0s)
-- Build files have been written to: /home/jfaraudo/papreca-main/build

So everything looks OK here. But then when doing the build I had this error:

jfaraudo@pcd31c:~/papreca-main/build$ cmake --build .
[  4%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/bond.cpp.o
In file included from /home/jfaraudo/papreca-main/source/libraries/PAPRECA/bond.cpp:23:
/home/jfaraudo/papreca-main/source/libraries/PAPRECA/bond.h:35:10: fatal error: pointers.h: No such file or directory
   35 | #include "pointers.h"
      |          ^~~~~~~~~~~~
compilation terminated.
gmake[2]: *** [CMakeFiles/PAPRECA_LIBRARY.dir/build.make:76: CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/bond.cpp.o] Error 1
gmake[1]: *** [CMakeFiles/Makefile2:85: CMakeFiles/PAPRECA_LIBRARY.dir/all] Error 2
gmake: *** [Makefile:91: all] Error 2

Any ideas?
