### 2024-06-05 #6714 @jfaraudo
@sntioudis I have tried a fresh install of LAMMPS and papreca on a new machine (with fedora Linux) . LAMMPS install was OK but I had errors with papreca. So I opened that issue:
https://github.com/sntioudis/papreca/issues/16
