### 2024-06-05 sntioudis/papreca#16 @jfaraudo
Oh!! Obviously this is the error, I missed the lammps folder . Sorry
