### 2024-06-05 sntioudis/papreca#16 @jfaraudo
Now I get the follwoing error.
I removed everything of course and I started over.

The preparation of CMake was OK:
jfaraudo@pcd31c:~/papreca-main/build$ cmake ../Installation/CMake -DLAMMPS_SRC_DIR=/home/jfaraudo/mylammps/src -DLAMMPS_LIB_DIR=/home/jfaraudo/mylammps/build
-- The C compiler identification is GNU 13.2.1
-- The CXX compiler identification is GNU 13.2.1
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /usr/bin/cc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Found LAMMPS library: /home/jfaraudo/mylammps/build/liblammps.a
-- Configuring additional packages...
CMake Warning at CMakeLists.txt:59 (message):
  PAPRECA does not use OpenMP, JPEG, and PNG packages.  However, the LAMMPS
  library uses them and therefore, we need to compile against those libraries
  to avoid linkage errors.  Depending on how YOUR LAMMPS library was built
  (i.e., based on the LAMMPS packages you enabled) you might need to enable
  additional libraries in the PAPRECA CMakeLists.txt file.

-- Found MPI_C: /usr/lib64/openmpi/lib/libmpi.so (found version "3.1")
-- Found MPI_CXX: /usr/lib64/openmpi/lib/libmpi_cxx.so (found version "3.1")
-- Found MPI: TRUE (found version "3.1")
-- Configuring done (2.4s)
-- Generating done (0.0s)
-- Build files have been written to: /home/jfaraudo/papreca-main/build

And now the build:

jfaraudo@pcd31c:~/papreca-main/build$ cmake --build .
[  4%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/bond.cpp.o
[  9%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/debug.cpp.o
[ 14%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/equilibration.cpp.o
[ 19%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/event.cpp.o
[ 23%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/event_detect.cpp.o
[ 28%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/event_execute.cpp.o
[ 33%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/event_list.cpp.o
[ 38%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/event_select.cpp.o
[ 42%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/export_files.cpp.o
[ 47%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/geometry_calc.cpp.o
[ 52%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/input_file.cpp.o
[ 57%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/lammps_wrappers.cpp.o
[ 61%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/mpi_wrappers.cpp.o
[ 66%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/papreca_config.cpp.o
[ 71%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/papreca_error.cpp.o
[ 76%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/rates_calc.cpp.o
[ 80%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/sim_clock.cpp.o
[ 85%] Building CXX object CMakeFiles/PAPRECA_LIBRARY.dir/home/jfaraudo/papreca-main/source/libraries/PAPRECA/utilities.cpp.o
[ 90%] Linking CXX static library libPAPRECA_LIBRARY.a
[ 90%] Built target PAPRECA_LIBRARY
[ 95%] Building CXX object CMakeFiles/papreca.dir/home/jfaraudo/papreca-main/source/papreca.cpp.o
[100%] Linking CXX executable papreca
/usr/bin/ld: libPAPRECA_LIBRARY.a(lammps_wrappers.cpp.o): in function `PAPRECA::initializeLMP(LAMMPS_NS::LAMMPS**)':
lammps_wrappers.cpp:(.text+0x143): undefined reference to `LAMMPS_NS::LAMMPS::LAMMPS(int, char**, ompi_communicator_t*)'
collect2: error: ld returned 1 exit status
gmake[2]: *** [CMakeFiles/papreca.dir/build.make:104: papreca] Error 1
gmake[1]: *** [CMakeFiles/Makefile2:111: CMakeFiles/papreca.dir/all] Error 2
gmake: *** [Makefile:91: all] Error 2

What do you think?
