### 2024-06-05 sntioudis/papreca#16 @sntioudis
Any chance you could also provide me with the cmake configuration output for LAMMPS as well? I need to check if there is a version mismatch between the MPI, and openMP libraries between LAMMPS and PAPRECA.
