### 2024-06-05 18ec005
Removed quite option from openMP
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
