+++ Installation/CMake/CMakeLists.txt
@@ -69,7 +69,7 @@ else()
 	message(FATAL_ERROR "MPI not found. Please make sure MPI is installed.")
 endif()
 
-find_package(OpenMP COMPONENTS CXX QUIET)
+find_package(OpenMP COMPONENTS CXX)
 if(OpenMP_FOUND)
 	target_link_libraries(papreca PRIVATE OpenMP::OpenMP_CXX)
 	target_link_libraries(PAPRECA_LIBRARY PRIVATE OpenMP::OpenMP_CXX)
