### 2024-06-05 sntioudis/papreca#16 @sntioudis
I tried to reproduce this error on my system but I failed. It looks like there is some sort of an MPI version mismatch (can happen if you have multiple versions of openMP or MPI in your system). It might also be that your LAMMPS installation did not finish properly or failed silently when trying to generate the LAMMPS library.

> /usr/bin/ld: libPAPRECA_LIBRARY.a(lammps_wrappers.cpp.o): in function PAPRECA::initializeLMP(LAMMPS_NS::LAMMPS**)': lammps_wrappers.cpp:(.text+0x143): undefined reference to LAMMPS_NS::LAMMPS::LAMMPS(int, char**, ompi_communicator_t*)'

This tells us that the linker could not find the constructor of the LAMMPS object.

Can you try to run this command and provide me with the output:

`nm -C /home/jfaraudo/mylammps/build/liblammps.a | grep LAMMPS_NS::LAMMPS::LAMMPS`

This might will let us know if there is a definition mismatch.
