### 2024-06-05 sntioudis/papreca#16 @jfaraudo
Here is
cmake -DPKG_MOLECULE=on -DPKG_RIGID=on -DPKG_QEQ=on -DPKG_REAXFF=on -DBUILD_LIB=on -DBUILD_SHARED_LIBS=off -DBUILD_STATIC_LIBS=on ../cmake
-- The CXX compiler identification is GNU 13.2.1
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Found Git: /usr/bin/git (found version "2.44.0")
-- Running check for auto-generated files from make-based build system
-- Checking for module 'mpi-cxx'
--   Package 'mpi-cxx', required by 'virtual:world', not found
-- Looking for C++ include omp.h
-- Looking for C++ include omp.h - found
-- Found OpenMP_CXX: -fopenmp (found version "4.5")
-- Found OpenMP: TRUE (found version "4.5") found components: CXX
-- Found GZIP: /usr/bin/gzip
-- Could NOT find FFMPEG (missing: FFMPEG_EXECUTABLE)
-- Looking for C++ include cmath
-- Looking for C++ include cmath - found
-- Generating style headers...
-- Generating package headers...
-- Generating lmpinstalledpkgs.h...
-- Found Python3: /usr/bin/python3.12 (found version "3.12.2") found components: Interpreter
-- Could NOT find ClangFormat (missing: ClangFormat_EXECUTABLE) (Required is at least version "8.0")
-- The following tools and libraries have been found and configured:
 * Git
 * OpenMP
 * Python3

-- <<< Build configuration >>>
   LAMMPS Version:   20240417 patch_17Apr2024
   Operating System: Linux Fedora 39
   CMake Version:    3.27.7
   Build type:       RelWithDebInfo
   Install path:     /home/jfaraudo/.local
   Generator:        Unix Makefiles using /usr/bin/gmake
-- Enabled packages: MOLECULE;QEQ;REAXFF;RIGID
-- <<< Compilers and Flags: >>>
-- C++ Compiler:     /usr/bin/c++
      Type:          GNU
      Version:       13.2.1
      C++ Flags:     -O2 -g -DNDEBUG
      Defines:       LAMMPS_SMALLBIG;LAMMPS_MEMALIGN=64;LAMMPS_OMP_COMPAT=4;LAMMPS_GZIP
-- <<< Linker flags: >>>
-- Executable name:  lmp
-- Static library flags:
-- Configuring done (4.9s)
-- Generating done (0.1s)
-- Build files have been written to: /home/jfaraudo/mylammps/build
