### 2024-06-05 sntioudis/papreca#16 @sntioudis
Thank you!

I think I see what the issue is now. For some reason, LAMMPS does not detect MPI at all.

This is what your LAMMPS cmake configuration output should look like:

```
cmake -DPKG_MOLECULE=on -DPKG_RIGID=on -DPKG_QEQ=on -DPKG_REAXFF=on -DBUILD_LIB=on -DBUILD_SHARED_LIBS=off -DBUILD_STATIC_LIBS=on ../cmake #Configure LAMMPS, build with some optional package, and enable static library building
-- The CXX compiler identification is GNU 11.4.0
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Found Git: /usr/bin/git (found version "2.34.1")
-- Running check for auto-generated files from make-based build system
-- Found MPI_CXX: /usr/lib/x86_64-linux-gnu/openmpi/lib/libmpi_cxx.so (found version "3.1")
-- Found MPI: TRUE (found version "3.1")
-- Looking for C++ include omp.h
-- Looking for C++ include omp.h - found
-- Found OpenMP_CXX: -fopenmp (found version "4.5")
-- Found OpenMP: TRUE (found version "4.5") found components: CXX
-- Found JPEG: /usr/lib/x86_64-linux-gnu/libjpeg.so (found version "80")
-- Found GZIP: /usr/bin/gzip
-- Could NOT find FFMPEG (missing: FFMPEG_EXECUTABLE)
-- Looking for C++ include cmath
-- Looking for C++ include cmath - found
-- Generating style headers...
-- Generating package headers...
-- Generating lmpinstalledpkgs.h...
-- Found Python3: /usr/bin/python3.10 (found version "3.10.12") found components: Interpreter
-- Could NOT find ClangFormat (missing: ClangFormat_EXECUTABLE) (Required is at least version "8.0")
-- The following tools and libraries have been found and configured:
 * Git
 * MPI
 * OpenMP
 * JPEG
 * Python3

-- <<< Build configuration >>>
   LAMMPS Version:   20240417 patch_17Apr2024
   Operating System: Linux Ubuntu" 22.04
   CMake Version:    3.22.1
   Build type:       RelWithDebInfo
   Install path:     /home/sn120/.local
   Generator:        Unix Makefiles using /usr/bin/gmake
-- Enabled packages: MOLECULE;QEQ;REAXFF;RIGID
-- <<< Compilers and Flags: >>>
-- C++ Compiler:     /usr/bin/c++
      Type:          GNU
      Version:       11.4.0
      C++ Flags:     -O2 -g -DNDEBUG
      Defines:       LAMMPS_SMALLBIG;LAMMPS_MEMALIGN=64;LAMMPS_OMP_COMPAT=4;LAMMPS_JPEG;LAMMPS_GZIP
-- <<< Linker flags: >>>
-- Executable name:  lmp
-- Static library flags:
-- <<< MPI flags >>>
-- MPI_defines:      MPICH_SKIP_MPICXX;OMPI_SKIP_MPICXX;_MPICC_H
-- MPI includes:     /usr/lib/x86_64-linux-gnu/openmpi/include;/usr/lib/x86_64-linux-gnu/openmpi/include/openmpi
-- MPI libraries:    /usr/lib/x86_64-linux-gnu/openmpi/lib/libmpi_cxx.so;/usr/lib/x86_64-linux-gnu/openmpi/lib/libmpi.so;
-- Configuring done
-- Generating done
-- Build files have been written to: /home/sn120/Desktop/LAMMPS/mylammps_17Apr2024/build
```

Notice that cmake informs me that *MPI has been correctly found and configured (the precise locations of MPI libraries are defined in the <<< MPI flags >>> section). Strangely enough, I can see that PAPRECA correctly identifies your MPI_CXX compiler:

> -- Found MPI_C: /usr/lib64/openmpi/lib/libmpi.so (found version "3.1")
-- Found MPI_CXX: /usr/lib64/openmpi/lib/libmpi_cxx.so (found version "3.1")
-- Found MPI: TRUE (found version "3.1")

Could you please uninstall LAMMPS and try to install it again by explicitly mentioning the name of the MPI compiler in your cmake command:

`cmake -DCMAKE_CXX_COMPILER=mpicxx -DPKG_MOLECULE=on -DPKG_RIGID=on -DPKG_QEQ=on -DPKG_REAXFF=on -DBUILD_LIB=on -DBUILD_SHARED_LIBS=off -DBUILD_STATIC_LIBS=on ../cmake`

Kindly let me know if your cmake configuration now correctly finds MPI before building.

If this works you can proceed by installing PAPRECA using a similar logic to generate your cmake configuration files:

`cmake ../Installation/CMake -DCMAKE_CXX_COMPILER=mpicxx -DLAMMPS_SRC_DIR=/home/jfaraudo/mylammps/src -DLAMMPS_LIB_DIR=/home/jfaraudo/mylammps/build`
