### 2024-06-05 sntioudis/papreca#16 @sntioudis
Just a small note here: It might be better to explicitly define the compiler by its location.

1) Find the location of your mpicxx by running: `which mpicxx`
2) Copy and paste that location to your DCMAKE_CXX_COMPILER variable (i.e., if your mpicxx is located here:/usr/bin/mpicxx your variable should be -DCMAKE_CXX_COMPILER=/usr/bin/mpicxx
