### 2024-06-05 sntioudis/papreca#16 @jfaraudo
Yeah, excellent suggestion .
I did as you suggested.
> which mpicxx
/usr/lib64/openmpi/bin/mpicxx
So then I did for LAMMPS:
cmake -DPKG_MOLECULE=on -DPKG_RIGID=on -DPKG_QEQ=on -DPKG_REAXFF=on -DBUILD_LIB=on -DBUILD_SHARED_LIBS=off -DBUILD_STATIC_LIBS=on -DCMAKE_CXX_COMPILER=/usr/lib64/openmpi/bin/mpicxx ../cmake
and for papreca:
 cmake ../Installation/CMake -DLAMMPS_SRC_DIR=/home/jfaraudo/mylammps/src -DLAMMPS_LIB_DIR=/home/jfaraudo/mylammps/build -DCMAKE_CXX_COMPILER=/usr/lib64/openmpi/bin/mpicxx

This solves the issue. I strongly suggest to add this to the building instructions, even this may sound trivial to you, it will be helpful to indicate that it is better to indicate the mpicxx compiler used.
