### 2024-06-05 sntioudis/papreca#16 @sntioudis
Glad to hear that the issue has been solved and happy to help.

> This solves the issue. I strongly suggest to add this to the building instructions, even this may sound trivial to you, it will be helpful to indicate that it is better to indicate the mpicxx compiler used.

This is a fair point and thank you for your feedback. I will add those instructions to the documentation shortly!
