### 2024-06-05 sntioudis/papreca#17 @jfaraudo
Example 1: comments and possible improvements
Here are my comments from testing the Examples provided with the software.
