### 2024-06-05 #6714 @jfaraudo
I am now testing the three examples provided with the Software.  I have a few comments about them, that I am posting in a new issue, starting from the Brownian motion elementary example.
https://github.com/sntioudis/papreca/issues/17
