### 2024-06-05 sntioudis/papreca#17 @jfaraudo
The first example is called **Brownian Self Diffusion**. It works as explained in the documentation. It is a fast executing example, easy to understand, so it is suitable as a first encounter with the software. But I have a few comments about it:
- Time step: I do not understand the reason for a timestep of 1.0e-100 seconds. It makes no sense to me. Especially, given the fact that the LAMMPS input file does not include any fixes moving the atoms, so there are no MD steps running here. The output of LAMMPS itself says that:
  " WARNING: No fixes with time integration, atoms won't move"
So Why are you including this number which has no physical/mathematical meaning?
- There is a group atom definition for atoms of type 2, but we have only atoms of type 1 as far as I see (in the LAMMPS input file, I mean the line with the command: "group fluid type 2" )
- The initial configuration has atom overlaps and produces a nonphysical very high energy output as seen here from the LAMMPS output:

  Step          Temp          E_pair         E_mol          TotEng         Press
         0   0              2.6601538e+15  0              2.6601538e+15  7.2961011e+14

This has to be corrected. The easiest way to do that is to add energy minimization steps before running the kMC simulation. In any case, you always need to start from a clean, energy minimized configuration without overlap so the generation of this configuration must be added previously to the kMC simulation. BTW, the atom overlaps can be easily seen in this snapshot in which I am showing the atoms (H atoms) with their Van der Waals size (whcih corresponds to the chosen LJ size in the LAMMPS forcefield).

Some minor issues:
a) the folder gets full of log files with data for the distributions. I suggest creating all these files in a subfolder so the processing of this data is separated from the general output . The access to the trajectory file, log file, and so one is the easier.
b) If you run the example as proposed in the documentation, a lot of relevant data is shown in the screen at a speed that cannot be read . It is better to redirect it to a log file. I suggest to include a redirection to a log file in the execution command proposed for the tutorial (I mean add something like  > lammpsoutput.log )
