### 2024-06-05 sntioudis/papreca#17 @jfaraudo
![snapshot](https://github.com/sntioudis/papreca/assets/30062302/cb67a623-6f3e-4551-99c4-d77b798669b9)
