### 2024-06-06 6c366b7
Added further installation instructions in documentation and changed Example scripts
