+++ documentation/examples.md
@@ -3,7 +3,7 @@
 \anchor examples
 
 > **Important note:**
-> To be able to run all the examples below, please ensure that you have built your LAMMPS library with the following packages: MOLECULE, RIGID, and QEQ. See \ref installation and the relevant [LAMMPS documentation page](https://docs.lammps.org/Build_package.html) for more information.
+> To be able to run all the examples below, please ensure that you have built your LAMMPS library with the following packages: EXTRA-DUMP, MOLECULE, RIGID, and QEQ. See \ref installation and the relevant [LAMMPS documentation page](https://docs.lammps.org/Build_package.html) for more information.
 
 \section brownian Your first PAPRECA simulation: Brownian Diffusion (random walk)
 
@@ -21,10 +21,6 @@ As far as the LAMMPS input file (i.e., in_kmc.lmp)  is concerned, note that:
 - This simulation is a pure kMC simulation. The choice of pair_style does not make any difference at all. We simply use a hybrid/overlay of
 two interatomic potentials to produce a half neighbor list (generated from the "lj/cut" potential) as well as a full list (obtained from the "zero" potential).
 - The pair_coeff 1 1 for the "lj/cut" potential is defined to set the sigma value (i.e., 2.84642 Angstroms). Such value is used to perform collision searches for the diffusion event (see \ref createDiff, \ref sigoptions, and \ref insigma).
-- A timestep of 1.0e-100 is deployed. This is done because we plan on running an MD stage every few kMC stages to output trajectory files. Of course, the
-MD stage will solely generate an "xyz" file and will have no effect on the system dynamics (i.e., the %PAPRECA run will be a pure off-lattice kMC simulation). This
-can be justified by the following fact: 1) the chosen timestep value (i.e., 1.0e-100 fs) is beyond the accuracy limits of double, and 2) the trajectory_duration (as in the %PAPRECA 
-input file below) is set to 1.
 
 For more information regarding the used LAMMPS commands consider visiting the relevant [documentation pages](https://docs.lammps.org/commands_list.html#).
 
@@ -36,7 +32,7 @@ When it comes to the %PAPRECA input file (i.e., in_kmc.ppc), note that:
 
 \subsection brownian_run Execution
 
-Execute the following command in the ./Examples/Brownian Self Diffusion/ folder to run the Brownian motion example:
+The easiest way to run %PAPRECA is by running one of the follwing commands in the directory where the input files are located:
 
 ```bash
 mpiexec ../build/papreca -in in_kmc.lmp in_kmc.ppc #modify papreca executable path if necessary
@@ -45,24 +41,43 @@ mpiexec ../build/papreca -in in_kmc.lmp in_kmc.ppc #modify papreca executable pa
 or
 
 ```bash
+mpiexec ../build/papreca -in in_kmc.lmp in_kmc.ppc > log.hybrid #modify papreca executable path if necessary
+```
+
+or
+
+```bash
 mpirun ../build/papreca -in in_kmc.lmp in_kmc.ppc #modify papreca executable path if necessary
 ```
 
-Of course, running the example requires that you have previously installed %PAPRECA (see \ref installation).
+or
 
-To run the Brownian motion example as an automated test use "test_brownian.sh" bash script in the example folder directory. Note that,
-you will have to modify the "papreca_dir" variable in your "test_brownian.sh" script so it points to the path of your %PAPRECA executable. If you wish to use "python" instead of "python3" you will also have to modify "test_brownian.sh". In any case, the Brownian test can be executed through the following command:
+```bash
+mpirun ../build/papreca -in in_kmc.lmp in_kmc.ppc > log.hybrid #modify papreca executable path if necessary
+```
+
+> **Important note 1:**
[…]
