### 2024-06-06 c6770db
Fixed minor typos in the examples section
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
