### 2024-06-06 sntioudis/papreca#17 @sntioudis
Thank you for your comments. Note that all the relevant changes have been commited to the “main” branch.

> Time step: I do not understand the reason for a timestep of 1.0e-100 seconds. It makes no sense to me. Especially, given the fact that the LAMMPS input file does not include any fixes moving the atoms, so there are no MD steps running here. The output of LAMMPS itself says that:
" WARNING: No fixes with time integration, atoms won't move"
So Why are you including this number which has no physical/mathematical meaning?

Fair point. Since there are no time integration fixes, atoms will not move anyway, regardless of the chosen timestep. Hence, I changed the timestep to 1 to avoid any confusion (even though it does not affect the results).

> There is a group atom definition for atoms of type 2, but we have only atoms of type 1 as far as I see (in the LAMMPS input file, I mean the line with the command: "group fluid type 2" )

For a PAPRECA simulation to run, both "group fluid" and "group frozen" commands have to be used in the PAPRECA input file (as mentioned in this part of the [documentation)](https://sntioudis.github.io/papreca/md_commands.html#flutypes). Both PAPRECA commands (i.e., "group fluid" and "group frozen") control the frozen and mobile atom groups, respectively, during the MD stage. Now, this is a "pure kMC example", and we want atoms of type 1 (i.e., our only type in the system) to remain frozen during the MD stage. Since there were no other atom types in the group, I randomly set atoms of type 2 to be “fluid” during the MD stage. This does not affect the results in any way (i.e., one could have selected group fluid type 99 without affecting the results). Nonetheless, it is mandatory to define a group fluid type and this is why this command was used in this way.

> BTW, the atom overlaps can be easily seen in this snapshot in which I am showing the atoms (H atoms) with their Van der Waals size (whcih corresponds to the chosen LJ size in the LAMMPS forcefield).

This example was designed such that the initial distribution of atoms was as close to a Diract Delta function as possible. The “denser” atoms are in the initial state, the better the Dirac Delta function is approximated. The example was inspired by [Einstein's Theory (see relevant section)](https://en.wikipedia.org/wiki/Brownian_motion), hence I wanted _"all particles to be located at the origin at t=0"_ (or at least as closely packed as possible) as mentioned in the label of the corresponding distributions graph.  Of course, this kind of initial structure leads to elevated potential energies (which do not affect the simulation results at all, since this is a “pure kMC” simulation and the interatomic potential is simply used as a placeholder to enable “collision tests”).

In any case, I made a few changes to the example to improve clarity. Firstly, I made the simulation box a bit bigger. Secondly, I included the “overlap” option to the create_atoms command. Hence, now LAMMPS does not insert atoms very closely to each other, preventing the potential energy from reaching unphysical values (even if this does not influence the results in any way). Finally, I introduced a minimization step that relaxes atoms based on the LJ potential. The changes above do not influence the results or the stability of the simulation but simply reduce the LAMMPS output value for the potential energy.

> a) the folder gets full of log files with data for the distributions. I suggest creating all these files in a subfolder so the processing of this data is separated from the general output . The access to the trajectory file, log file, and so one is then easier.

The Brownian test should be executed as an automated test. I should have mentioned that clearly in the documentation, hence, this is my bad. I made some changes to the documentation to encourage users to run this as an automated test.  In any case, when you run the Brownian example as an automated test, and at the end of your simulation, the test_brownian.sh will "tidy everything up" by creating a new folder called results and moving all your log output files over there. This was necessary since PAPRECA creates the distribution files at the parent directory, by default. The trajectory file(s), the log.lammps file, papreca.log, execTimes.log, and the output image (“distributions.jpg”) are not moved in the results folder but remain in the parent directory for easier access. In any case, I decreased the export output of the elemental distribution files and now less files are generated in the folder.

> b) If you run the example as proposed in the documentation, a lot of relevant data is shown in the screen at a speed that cannot be read . It is better to redirect it to a log file. I suggest to include a redirection to a log file in the execution command proposed for the tutorial (I mean add something like > lammpsoutput.log )

LAMMPS automatically generates a file called log.lammps with all info related to the MD stage. Another file called papreca.log is generated to output some of the information related to the kMC stage. Therefore, almost all simulation info is already printed in certain files of the parent directory. I refrained from suggesting the redirection of the output to a new file for a couple off reasons. Firstly, it generates very big files. Secondly, I have observed certain performance issues when redirecting text files produced from MPI scripts.

Regardless, I am aware that some very specific information regarding the kMC events (e.g., vacancy position) are not printed in the papreca.log file.  Exporting such data would require communication between the various MPI processes and would probably drop the efficiency of the code. Consequently, I added a note in the documentation that encourages people to output the screen data in a separate file (but warns them that this can generally lower the computational efficiency of the code).

> c) the xyz trajectory file is quite large. Why not saving the data as a compressed dcd file? I usually save the initial configuration as pdb (uncompressed, human-readable) and then the trajectory in a binary format such as dcd which saves a lot of space.

Dumping files in dcd format can only be done if LAMMPS is compiled with the EXTRA-DUMP package. I made the relevant changes to the example but keep in mind that it only runs if you re-install LAMMPS with the EXTRA-DUMP package.
