### 2024-06-06 #6714 @jfaraudo
> > @sntioudis I have tried a fresh install of LAMMPS and papreca on a new machine (with fedora Linux) . LAMMPS install was OK but I had errors with papreca. So I opened that issue: [sntioudis/papreca#16](https://github.com/sntioudis/papreca/issues/16)
>
> Installed and working, now I'll proceed to test the examples.
>
> As a suggestion for the documentation, I think it is better to add as recommendation the procedure that solved my compilation problems. Locate first your mpicxx using which mpicxx and then use this path in the cmake with the option -DCMAKE_CXX_COMPILER=/path/to/your/mpicxx (in my case /usr/lib64/openmpi/bin/mpicxx )
>
> Once this is included in the documentation, the issue #16 can be closed [sntioudis/papreca#16](https://github.com/sntioudis/papreca/issues/16)

@sntioudis I see that you closed this issue but the comment about cxx_compiler is still not included in the documentation as suggested (maybe it is uploading or something?)
