### 2024-06-06 sntioudis/papreca#17 @jfaraudo
Thank you @sntioudis for your detailed responses. Let me read it asap. BTW, concerning the issue with big data files and dcd files possibility, you are right. I always store my data in dcd format but it is true that in LAMMPS this requires an additional flag in compilation, so you can forget about my comment about dcd, it makes it more complicated for users that do no use this format by default.
