### 2024-06-06 c90241e
Added note regarding VMD and OVITO
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
