### 2024-06-06 a54fc61
Update examples.md
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
