### 2024-06-06 #6714 @srmnitc
> @srmnitc I finished executing the examples and revising the documentation. Overall, I like the work , @sntioudis you did a great job. I have a few comments and suggestions, I'll post all of them asap.

Thanks a lot for your efforts @jfaraudo !
