### 2024-06-06 sntioudis/papreca#17 @jfaraudo
Thank you for your efforts. I agree that none of the issues I mentioned affect the results, but I think the effort of tackling these is worthwhile since a couple of them were a bit odd. I think that now everything is much clear.
