### 2024-06-06 sntioudis/papreca#17 @sntioudis
@jfaraudo I agree. Example #1 is definitely clearer now, so your input is very much appreciated!
