### 2024-06-06 sntioudis/papreca#19 @jfaraudo
In general, the documentation is clear and covers all relevant aspects. I have only two minor comments:
- For me, the flowchart in the section "Essential information (See before you start)"  athttps://sntioudis.github.io/papreca/md_essentials.html  is a bit confusing. I agree that all the information is correct. But at a first look , for an expert in MD like me, it looks pretty strange to see, in the MD part,  the sequence of minimization - MD run - and minimization again, because it is pretty unusual in the MD community. I understood that flow when looking at your paper  Computational Materials Science, vol. 229, 2023. The flow in Figure 1 of that paper is much clear to me. It is essentially the same, but it contains a schematic with a PES and the exploration over the PES implied by these steps. This makes it more clear to me. Also, the general organization of the flowchart is better. I suggest considering an update of this flowchart recycling elements of (or based on) that previous figure.
- I do not understand the reason of the comment about LJ units in the section "PAPRECA units". I agree that there is no simple way of converting LJ units to real time, that's clear. But imagine that we select LJ units in LAMMPS. Then, if PAPRECA is also working in LJ units (if I understand correctly, PAPRECA uses the same units as LAMMPS), why not providing the time of MC, MD and total time in LJ units (no need of conversion to seconds)?
