### 2024-06-06 #6714 @jfaraudo
@srmnitc @sntioudis I closed the issues related to the Examples. I am fully satisfied with the corrections and updates made by @sntioudis to the Examples. I especially like the fact that Example 3 corresponds to an actual research paper. The selection of examples is great. Example 1 teaches you how to work with this software. Example 2 shows a classical example that can be tested with analytical solutions.  And finally example 3 shows a Research paper result.
I have a few minor comments about the paper itself and the documentation that are posted in the following issues:
https://github.com/sntioudis/papreca/issues/15
https://github.com/sntioudis/papreca/issues/19
I think that, after consideration of these comments, the work can be published.
