### 2024-06-06 sntioudis/papreca#15 @jfaraudo
I have another comment about the paper. After reading the paper, it was not obvious to me why kMC and MD were combined in this way. After reading your previous the paper in Computational Materials Science, vol. 229, 2023, that was obvious to me. You mention clearly in the pevious paper that the mission of the MD is only to take into account high frequency / short time effects such as molecular vibration, small changes in bond distances and molecular flexibility and thinks like that, whereas the main kinetics of the effects to be studied were considered using kMC from "frozen" MD configurations. I think that a short explanation in this line, like the one in your paper (for ex. in page 3 below the flow chart), will be a great addition to the Statement of need section of the JOSS paper.
