### 2024-06-06 sntioudis/papreca#4 @jfaraudo
I also agree with @liamhuber , it will be great to have this sofware available with conda. It makes life much easier to users. I am also thinking about students, in my undergrad or summer courses it is usually not a big deal to ask people to install something from conda but doing more than that could be complicated. It is not urgent, but it will be great one day to have this possibility.
