### 2024-06-06 sntioudis/papreca#19 @jfaraudo
All of these sounds good to me. I do not think that the "LJ community" so to speak is really working a lot on kMC these days. So I think it is not a big priority to implement a LJ units converter. But I think it is worth explaining in more detail that PAPRECA uses "real" units so all other LAMMPS units options are handled by internal conversion to "real" at PAPRECA.
