### 2024-06-06 9b30bb2
Added note on kMC stage units
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
