### 2024-06-07 sntioudis/papreca#15 @sntioudis
@jfaraudo
> The authors mention at several points the generality of the method, and in the statement of need, they include examples such as protein folding or enzyme reactions that require consideration of solvent, in order to be simulated properly. I am not completelly sure that the present method works efficiently in presence of solvent. In my experience, explicit solvent makes MC methods quite inefficient (pls note that I have still to run the examples, so I admit that I can be wrong). For me it is clear that in situations such as solid state simulations , molecules in contact with solids and thinks like this the method is perfect but I think that a comment whether this method is recommended or not in simulations with explicit solvent is in order.

I took a couple of days to think about this and I believe that the presence of solvent does not directly affect the performance of kinetic Monte Carlo codes. Specifically, the kMC stage of PAPRECA definitely becomes more expensive if:

1) The number of atoms increases. If this happens then more event searches have to be performed.
2) The number of predefined events increases. In this case, once again, the number of event searches rises.
3) The number of atoms included in the neighbors lists elevates. In essence, this would mean that the system is "denser" which tranlates to more expensive event searches (e.g., when we are looking for bond-pairs or checking for collisions in diffusion or adsorption events).

I reckon that the presence of solvent can make the kMC stage of the simulation more expensive, but indirectly. In other words, explicitly including the solvent introduces atoms to the system and potentially the neighbors lists. Nevertheless, simulating a very dense solid/vapour interface might become computationally demanding for the exact same reasons (i.e., if the film is too "dense" or the interface/surface area is too large).

I admit that the MD stage of a hybrid PAPRECA run (i.e., kMC+MD) should be impacted in some way by the presence of solvent. Nonetheless, this is a LAMMPS topic and not necessarily something relevant with PAPRECA or kMC simulations including solvents.

I performed the following simulations as a sanity check.

1) Unsolvated tri-m-cresyl molecules in a box (5452 atoms).

![tcp](https://github.com/sntioudis/papreca/assets/42137665/42bf0054-0126-4a89-ae1d-78e090b1a2ad)

```
Final stats: Total KMC walltime= 	 	11.27842874 sec (0.60280958%)
	 	 	 Total MD walltime= 	 	1859.69859454 sec (99.39719042%)
	 	 	 Total HYBRID KMC/MD time=  1870.97702328 sec
```
[TCP_unsolvated.zip](https://github.com/user-attachments/files/15728404/TCP_unsolvated.zip)

2) Mixture of tri-m-cresyl molecules and toluene (solvent) molecules (5440 atoms).

![tcp_toluene](https://github.com/sntioudis/papreca/assets/42137665/083354d6-1319-4952-8715-5f47f008c7f1)

```
Final stats: Total KMC walltime= 	 	4.20634362 sec (1.52989671%)
	 	 	 Total MD walltime= 	 	270.73663669 sec (98.47010329%)
	 	 	 Total HYBRID KMC/MD time=  274.94298030 sec
```
[TCP_TOLUENE.zip](https://github.com/user-attachments/files/15728434/TCP_TOLUENE.zip)

Note that both systems were simulated using the exact same predefined events, simulation box settings, and interatomic potential. Even though both simulations had (almost)  equal number of atoms, we observe that the kMC stage of the unsolvated system was about 3 times faster than the relevant stage of the solvated system. To the best of my understanding, this can be attributed to the fact that the unsolvated system formed agglomerates, elevating the number of atoms in the neighbors lists. In turn, this led to a rise in the number of "neighboring atom searches" of the bond formation events defined in the system (i.e., P-O, O-O, and C-O). Conclusively, explicitely including the solvent atoms in a system does not guarantee that the computational cost would rise.

I am attaching compressed files with the input files, in case you want to run this.
