### 2024-06-07 sntioudis/papreca#15 @jfaraudo
Thank you for considering these tests!. Traditionally MC is not very popular with people working with systems in solvent because for most solvents (especially for water) it is very hard to find "cavities" able to allocate a molecule jumping from one  place to another one. I mean, if you have for example a molecule diffusing from one place to another in solvent, to simulate this motion with MC you need a low energy place to put this molecule without overlaping the solvent. Overlaping the solvent will have an incredibly large energy cost in a MC move , so it makes the probability of the MC move almost zero. In your case, with the kMC you will have for sure overlaps with solvent molecules when you move your solute so then you need to minimize again energy (OK, something that you do in your PAPRECA + LAMMPS workflow) to move the overlaped solvent molecules somewhere. That step could be also hard to converge if the liquid is dense (for example, water, which is quite dense, even more than its solid phase, you know). The test you show are interesting. Organic solvents as you considered are also very important. Since these systems are very different from anything alse considered in PAPRECA, my suggestion is to add these somewhere as examples of runs with solvent or something like that. No need for extensive documentation (not at the level of the examples 1,2 and 3), I think it can be provided just as a proof of concept that the methods can be used in opresence of solvent. For people working with solvents it will be of interest to see that.
