### 2024-06-07 sntioudis/papreca#15 @sntioudis
Thank you for your input and I certainly agree with all of the points made. To the best of my understanding, the performance of the system would really depend on the system. What worked ok with the TCP+toluene example might not perform so well in a water based system (potentially due to the elevated cost of energy minimization as you correctly mention.

`my suggestion is to add these somewhere as examples of runs with solvent or something like that. No need for extensive documentation (not at the level of the examples 1,2 and 3), I think it can be provided just as a proof of concept that the methods can be used in the presence of solvent. For people working with solvents it will be of interest to see that.`

Fair point. I will add this in the documentaiton.
