### 2024-06-07 16caf06
Updates in documentation: added solvents description and images and modified general workflow
