### 2024-06-07 sntioudis/papreca#19 @sntioudis
1) I updated the general PAPRECA flowchart in the [Essentials section](https://sntioudis.github.io/papreca/md_essentials.html#flowchart). I realized that the General flowchart in the documentation with the two minimization steps (i.e., one before and one after executing the MD trajectory) was referring to the phosphates examples (and to our [previous paper]). PAPRECA is a more general software, in the sense that the user has complete freedom over the MD simulation options. For example, someone might choose to perform only one minimization (i.e., before the MD trajectory), or none (as exhibited in some of the examples). Therefore, I removed the specific steps and the flowchart now depicts a general MD stage (without further description).

2) A more detailed discussion regarding the units of PAPRECA is now in the [essentials](https://sntioudis.github.io/papreca/md_essentials.html#units) section.
