### 2024-06-07 sntioudis/papreca#15 @sntioudis
@jfaraudo I included a [solvents](https://sntioudis.github.io/papreca/md_examples.html#solvents) example in the documentation and uploaded the relevant files in the repository.

I will be addressing the rest of the paper issues in a bit!
