### 2024-06-07 02d4ccb
Update paper.md
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
