+++ paper/paper.md
@@ -46,9 +46,11 @@ EON [@Chill2014May] is the only identified off-lattice package distributed under
 
 To the best of our knowledge, a completely lattice-free kMC code with predefined table of events is currently unavailable in open-source repositories. ```PAPRECA``` aims to fill that gap by providing the scientific community with a general and easy-to-use solution for performing long-timescale atomistic simulations on complex materials science, chemistry, physics, biology, and engineering problems involving amorphous materials or non-crystalline systems.
 
-```PAPRECA``` (in its initial version) can capture four distinct classes of predefined transition events: 1) reactions (bonding and scission), 2) depositions (of molecules and atoms), 3) diffusion hops, and 4) monoatomic desorptions. Virtually any system whose temporal evolution can be described by these atomic-scale processes can be effectively simulated by ```PAPRECA```. Example applications of ```PAPRECA``` include but are not limited to: adsorption/desorption on catalytic surfaces, amorphous thin films (e.g., phosphate films, solid electrolyte interphases, oxide layers), and modeling self-diffusion of gases. Furthermore, ```PAPRECA``` allows for the extension of the source code to include other classes of transition events (e.g., reaction chains). 
+```PAPRECA``` (in its initial version) is a parallelized software (uses the MPI protocol) that can capture four distinct classes of predefined transition events: 1) reactions (bonding and scission), 2) depositions (of molecules and atoms), 3) diffusion hops, and 4) monoatomic desorptions. Virtually any system whose temporal evolution can be described by these atomic-scale processes can be effectively simulated by ```PAPRECA```. Example applications of ```PAPRECA``` include but are not limited to: adsorption/desorption on catalytic surfaces, amorphous thin films (e.g., phosphate films, solid electrolyte interphases, oxide layers), and modeling self-diffusion of gases. Furthermore, ```PAPRECA``` allows for the extension of the source code to include other classes of transition events (e.g., reaction chains). 
 
-Finally, ```PAPRECA``` is a parallelized code (uses the MPI protocol) that couples an off-lattice kMC solver with a MD solver (LAMMPS [@Thompson2022Feb]) to enable hybrid off-lattice kMC/MD simulations with a predefined table of events. Deploying a dual-solver approach can be beneficial for capturing both fast atomic-scale processes (whose frequency is comparable to that of atomic vibrations) through the MD stage as well as elementary events of elevated activation energies via the off-lattice kMC stage.
+For accurate simulations of atomistic fidelity, it becomes necessary to model elementary steps of high frequency ($10^{13}$-$10^{13}$ Hz and above [@VanSwygenhoven2006May]) such as vibrations of particles. These events cannot be explicitly included in the table of predefined events as they would dominate the system. This can be explained by considering that the higher the rate of a predefined elementary step, the greater its kMC selection probability [Fichthorn1991Jul]. To circumvent this issue, ``PAPRECA``` couples an off-lattice kMC solver with a MD solver (LAMMPS [@Thompson2022Feb]) to enable hybrid off-lattice kMC/MD simulations. Effectively, atomistic processes of elevated activation energies are captured via the off-lattice kMC stage, while fast atomic-scale processes are treated by the MD stage.
+
[…]
