### 2024-06-07 6d7b176
Update paper.bib
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
