+++ paper/paper.bib
@@ -416,3 +416,83 @@
 	publisher = {Frontiers},
 	doi = {10.3389/fchem.2019.00202}
 }
+
+@article{VanSwygenhoven2006May,
+	author = {Van Swygenhoven, Helena and Weertman, Julia R.},
+	title = {{Deformation in nanocrystalline metals}},
+	journal = {Materials Today},
+	volume = {9},
+	number = {5},
+	pages = {24--31},
+	year = {2006},
+	month = may,
+	issn = {1369-7021},
+	publisher = {Elsevier},
+	doi = {10.1016/S1369-7021(06)71494-8}
+}
+
+@article{Barhaghi2022Aug,
+	author = {Barhaghi, Mohammad Soroush and Crawford, Brad and Schwing, Gregory and Hardy, David J. and Stone, John E. and Schwiebert, Loren and Potoff, Jeffrey and Tajkhorshid, Emad},
+	title = {{py-MCMD: Python Software for Performing Hybrid Monte Carlo/Molecular Dynamics Simulations with GOMC and NAMD}},
+	journal = {Journal of Chemical Theory and Computation},
+	volume = {18},
+	number = {8},
+	pages = {4983--4994},
+	year = {2022},
+	month = aug,
+	issn = {1549-9618},
+	publisher = {American Chemical Society},
+	doi = {10.1021/acs.jctc.1c00911}
+}
+
+@article{Phillips2020Jul,
+	author = {Phillips, James C. and Hardy, David J. and Maia, Julio D. C. and Stone, John E. and Ribeiro, Jo{\ifmmode\tilde{a}\else\~{a}\fi}o V. and Bernardi, Rafael C. and Buch, Ronak and Fiorin, Giacomo and H{\ifmmode\acute{e}\else\'{e}\fi}nin, J{\ifmmode\acute{e}\else\'{e}\fi}r{\ifmmode\hat{o}\else\^{o}\fi}me and Jiang, Wei and McGreevy, Ryan and Melo, Marcelo C. R. and Radak, Brian K. and Skeel, Robert D. and Singharoy, Abhishek and Wang, Yi and Roux, Beno{\ifmmode\hat{\imath}\else\^{\i}\fi}t and Aksimentiev, Aleksei and Luthey-Schulten, Zaida and Kal{\ifmmode\acute{e}\else\'{e}\fi}, Laxmikant V. and Schulten, Klaus and Chipot, Christophe and Tajkhorshid, Emad},
+	title = {{Scalable molecular dynamics on CPU and GPU architectures with NAMD}},
+	journal = {Journal of Chemical Physics},
+	volume = {153},
+	number = {4},
+	year = {2020},
+	month = jul,
+	issn = {0021-9606},
+	publisher = {AIP Publishing},
+	doi = {10.1063/5.0014475}
+}
+
+@article{Nejahi2019,
+title = {{GOMC: GPU Optimized Monte Carlo for the simulation of phase equilibria and physical properties of complex fluids}},
+journal = {SoftwareX},
+volume = {9},
+pages = {20-27},
+year = {2019},
+issn = {2352-7110},
+doi = {https://doi.org/10.1016/j.softx.2018.11.005},
+url = {https://www.sciencedirect.com/science/article/pii/S2352711018301171},
+author = {Nejahi, Younes and Barhaghi, Mohammad Soroush and Mick, Jason and Jackman,Brock and Rushaidat, Kamel and Li, Yuanzhe and Schwiebert, Loren and Potoff, Jeffrey},
+}
+
+@article{Martin1999May,
+	author = {Martin, Marcus G. and Siepmann, J. Ilja},
[…]
