### 2024-06-07 d8a5a1e
Update paper.md
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
