+++ paper/paper.md
@@ -50,7 +50,7 @@ To the best of our knowledge, a completely lattice-free kMC code with predefined
 
 For accurate simulations of atomistic fidelity, it becomes necessary to model elementary steps of high frequency ($10^{13}$-$10^{13}$ Hz and above [@VanSwygenhoven2006May]) such as vibrations of particles. These events cannot be explicitly included in the table of predefined events as they would dominate the system. This can be explained by considering that the higher the rate of a predefined elementary step, the greater its kMC selection probability [Fichthorn1991Jul]. To circumvent this issue, ``PAPRECA``` couples an off-lattice kMC solver with a MD solver (LAMMPS [@Thompson2022Feb]) to enable hybrid off-lattice kMC/MD simulations. Effectively, atomistic processes of elevated activation energies are captured via the off-lattice kMC stage, while fast atomic-scale processes are treated by the MD stage.
 
-py-MCMD [@Barhaghi2022Aug] is a different hybrid MC/MD workflow available in open-source repositories. py-MCMD is a Python-based communication interface between the MC software GOMC[@Nejahi2019] and the MD code NAMD[@Phillips2020Jul]. The central difference between ```PAPRECA``` and py-MCMD is that the former implements a kMC approach, while the latter utilizes a Metropolis Monte Carlo (MMC) scheme. Effectively, in the kMC stage of ```PAPRECA``` predefined event probabilities are weighted based on their rates and an elementary process is selected and executed to overcome a high energy barrier. On the other hand, during the MC phase of py-MCMD, trial moves (e.g., rigid-body displacements/rotations, intra-box swaps [@Martin1999May]) are attempted and accepted or rejected based on the Metropolis acceptance criterion [@Chen2015Jan].
+py-MCMD [@Barhaghi2022Aug] is a different hybrid MC/MD workflow available in open-source repositories. py-MCMD is a Python-based communication interface between the MC software GOMC[@Nejahi2019] and the MD code NAMD[@Phillips2020Jul]. The central difference between ```PAPRECA``` and py-MCMD is that the former implements a kMC approach, while the latter utilizes a Metropolis Monte Carlo (MMC) scheme. Effectively, in the kMC stage of ```PAPRECA``` predefined event probabilities are calculated based on their rates and an elementary process is selected and executed to overcome a high energy barrier. On the other hand, during the MC phase of py-MCMD, trial moves (e.g., rigid-body displacements/rotations, intra-box swaps [@Martin1999May]) are attempted and accepted or rejected based on the Metropolis acceptance criterion [@Chen2015Jan].
 
 # Scalability of PAPRECA
 
