### 2024-06-07 392a456
Update paper.md
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
