+++ paper/paper.md
@@ -48,7 +48,7 @@ To the best of our knowledge, a completely lattice-free kMC code with predefined
 
 ```PAPRECA``` (in its initial version) is a parallelized software (uses the MPI protocol) that can capture four distinct classes of predefined transition events: 1) reactions (bonding and scission), 2) depositions (of molecules and atoms), 3) diffusion hops, and 4) monoatomic desorptions. Virtually any system whose temporal evolution can be described by these atomic-scale processes can be effectively simulated by ```PAPRECA```. Example applications of ```PAPRECA``` include but are not limited to: adsorption/desorption on catalytic surfaces, amorphous thin films (e.g., phosphate films, solid electrolyte interphases, oxide layers), and modeling self-diffusion of gases. Furthermore, ```PAPRECA``` allows for the extension of the source code to include other classes of transition events (e.g., reaction chains). 
 
-For accurate simulations of molecular-level resolution, it becomes necessary to model elementary steps of high frequency ($10^{13}$-$10^{13}$ Hz and above [@VanSwygenhoven2006May]) such atomic vibrations. These events cannot be explicitly included in the table of predefined events as they would dominate the system. This can be explained by considering that the higher the rate of a predefined elementary step, the greater its kMC selection probability [@Fichthorn1991Jul]. To circumvent this issue, ``PAPRECA``` couples an off-lattice kMC solver with a MD solver (LAMMPS [@Thompson2022Feb]) to enable hybrid off-lattice kMC/MD simulations. Effectively, atomistic processes of elevated activation energies are captured via the off-lattice kMC stage, while fast atomic-scale processes are treated by the MD stage.
+For accurate simulations of molecular-level resolution, it becomes necessary to model elementary steps of high frequency ($10^{13}$-$10^{13}$ Hz and above [@VanSwygenhoven2006May]) such as atomic vibrations. These events cannot be explicitly included in the table of predefined events because they would dominate the system, thus preventing the simulation from reaching timescales beyond the limits of MD. This can be explained by considering that the higher the rate of a predefined elementary step, the greater its kMC selection probability [@Fichthorn1991Jul]. To circumvent this issue, ``PAPRECA``` couples an off-lattice kMC solver with a MD solver (LAMMPS [@Thompson2022Feb]) to enable hybrid off-lattice kMC/MD simulations. Effectively, atomistic processes of elevated activation energies are captured via the off-lattice kMC stage, while fast atomic-scale processes are treated by the MD stage.
 
 py-MCMD [@Barhaghi2022Aug] is a different hybrid MC/MD workflow available in open-source repositories. py-MCMD is a Python-based communication interface between the MC software GOMC[@Nejahi2019] and the MD code NAMD[@Phillips2020Jul]. The central difference between ```PAPRECA``` and py-MCMD is that the former implements a kMC approach, while the latter utilizes a Metropolis Monte Carlo (MMC) scheme. Effectively, in the kMC stage of ```PAPRECA``` predefined event probabilities are calculated based on their rates and an elementary process is selected and executed to overcome a high energy barrier. On the other hand, during the MC phase of py-MCMD, trial moves (e.g., rigid-body displacements/rotations, intra-box swaps [@Martin1999May]) are attempted and accepted or rejected based on the Metropolis acceptance criterion [@Chen2015Jan].
 
