### 2024-06-10 c802bdd
Minor changes in documentation
