### 2024-06-10 #6714 @sntioudis
> I have a few minor comments about the paper itself and the documentation that are posted in the following issues:
https://github.com/sntioudis/papreca/issues/15
https://github.com/sntioudis/papreca/issues/19
I think that, after consideration of these comments, the work can be published

Hi @srmnitc. The issues above have been resolved.
