### 2024-06-10 #6714 @sntioudis
@srmnitc I am replying in a separate comment because I was unable to tick the boxes in the checklist.

>  Double check authors and affiliations (including ORCIDs)

Done.

>  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Version: v1.0.1
Github link: https://github.com/sntioudis/papreca/releases/tag/v1.0.1

>  Archive the release on Zenodo/figshare/etc and post the DOI here.

v1.0.1 archived in Zenodo: https://doi.org/10.5281/zenodo.11550493

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Done.

>  Make sure that the license listed for the archive is the same as the software license.

Done.
