### 2024-06-10 c7d3fbc
Update paper.bib
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
