+++ paper/paper.bib
@@ -465,7 +465,7 @@ volume = {9},
 pages = {20-27},
 year = {2019},
 issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2018.11.005},
+doi = {10.1016/j.softx.2018.11.005},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711018301171},
 author = {Nejahi, Younes and Barhaghi, Mohammad Soroush and Mick, Jason and Jackman,Brock and Rushaidat, Kamel and Li, Yuanzhe and Schwiebert, Loren and Potoff, Jeffrey},
 }
