### 2024-06-10 ea06620
Update paper.md
Signed-off-by: Sarath Menon <mail@sarathmenon.me>
