+++ paper/paper.md
@@ -33,14 +33,14 @@ bibliography: paper.bib
 
 # Summary
 
-Kinetic Monte Carlo (kMC) is an atomistic and stochastic simulation technique that captures the temporal evolution of various systems in materials science, chemistry, physics, biology, and engineering. Several open-source kMC packages are currently distributed online. Nevertheless, such implementations are typically lattice-based and are mostly designed to study ordered, crystalline materials. In this work, we present PArallel PREdefined CAtalog (```PAPRECA```), an easy-to-use and completely lattice-free open-source kMC software suitable for simulations on amorphous materials or systems characterized by a low degree of crystallinity. ```PAPRECA``` is a parallel C++ software using the Message Passing Interface (MPI) protocol and coupled with the Large-scale Atomic/Molecular Massively Parallel Simulator (LAMMPS) [@Thompson2022Feb] to enable pure kMC runs as well as hybrid kMC/MD simulations.
+Kinetic Monte Carlo (kMC) is an atomistic and stochastic simulation technique that captures the temporal evolution of various systems in materials science, chemistry, physics, biology, and engineering. Several open-source kMC packages are currently distributed online. Nevertheless, such implementations are typically lattice-based and are mostly designed to study ordered, crystalline materials. In this work, we present PArallel PREdefined CAtalog (```PAPRECA```), an easy-to-use and completely lattice-free open-source kMC software suitable for simulations on amorphous materials or systems characterized by a low degree of crystallinity. ```PAPRECA``` is a parallel C++ software using the Message Passing Interface (MPI) protocol and coupled with the Large-scale Atomic/Molecular Massively Parallel Simulator (LAMMPS) [@Thompson2022Feb] to enable pure kMC runs as well as hybrid kMC/Molecular Dynamics (MD) simulations.
 
 # Statement of need
 
-KMC models have been deployed to investigate non-equilibrium dynamics and properties of thin films [@Ntioudis2023Oct], nanoparticles [@Turner2016Apr], quantum dots [@Zhu2007May], semiconductors [@vanderKaap2016Feb], catalysts [@Stamatakis2012Dec], energy-storage devices [@Abbott2022Feb], interstellar grain chemistry [@Cuppen2013Dec], protein folding [@Makarov2001Jun], and enzyme reactions [@Slepoy2008May]. Overall, kMC techniques are, on one hand, less accurate, but on the other hand, more computationally efficient than molecular dynamics (MD). This is justified by the fact that kMC does not describe atomic vibrations explicitly but evolves the system through discrete elementary processes (e.g., diffusions, depositions, reactions etc.) [@Gillespie1976Dec ; @Fichthorn1991Jul]. In any case, the efficiency of kMC models unlocks the possibility for long-timescale simulations with molecular-level resolution beyond the timescales accessible to MD. Note that efficient MD algorithms can be used to simulate systems with $10^{4}$-$10^{6}$ atoms from ps to $\mu$s [@Thompson2022Feb]).
[…]
