### 2024-06-10 sntioudis/papreca#22 @srmnitc
Software in bib file
This is part of review at https://github.com/openjournals/joss-reviews/issues/6714

The misc citations in bib file also includes journal and publisher. These fields should be ideally removed. Another option would be to use inline links directly instead of citations.
