### 2024-06-10 #6714 @srmnitc
> @sntioudis I have opened a pull request with some small changes to the paper. Feel free to implement them yourself as you see fit, or merge it in. I have also opened an issue regarding the bib files, it would be great if you could please take a look.

After these final issues are done, I will finish the final steps from my side.
