### 2024-06-11 328f3ec
Removed Journal and publisher from paper.bib
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
