+++ paper/paper.bib
@@ -2,8 +2,6 @@
   author = {Ntioudis, Stavros and Ewen, James P. and Dini, Daniele and Turner, C. Heath},
   title = {papreca},
   year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/sntioudis/papreca}
 }
 
@@ -281,8 +279,6 @@
   author = {Reuter, Karsten and Savara, Aditya, and Hoffmann, Max J. and Matera, Sebastian and Lorenzi, Juan M. and Andersen, Mie and Bajdich, Michal and Garhammer, Andreas.},
   title = {kmcos: Kinetic Monte Carlo of Systems},
   year = {2020},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/kmcos/kmcos}
 }
 
@@ -290,8 +286,6 @@
   author = {Pasumarthi, Viswanath},
   title = {{Python-based Charge Dynamics (PyCD)}},
   year = {2017},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/vpasumarthi/PyCD/}
 }
 
@@ -326,8 +320,6 @@
   author = {Grabowski, Frederic and Kochanczyk, Marek},
   title = {{VIS-A-VIS: agent-based simulator of viral infection spread and viral infection self-containment in a monolayer of cell}},
   year = {2022},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/grfrederic/visavis}
 }
 
@@ -335,8 +327,6 @@
   author = {Helleboid, Remi},
   title = {{MulSKIPS: A Kinetic Monte Carlo super-Lattice code}},
   year = {2021},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/MulSKIPS/MulSKIPS}
 }
 
@@ -344,8 +334,6 @@
   author = {Jansson, Ville},
   title = {{Kimocs - Kinetic Monte Carlo for Surfaces}},
   year = {2016},
-  publisher = {GitLab},
-  journal = {GitLab repository},
   url = {https://gitlab.com/vjansson/Kimocs}
 }
 
@@ -353,8 +341,6 @@
   author = {Nandipati, Giridhar},
   title = {{KSOME: Kinetic Simulations of Microstructural Evolution}},
   year = {2021},
-  publisher = {GitLab},
-  journal = {GitLab repository},
   url = {https://gitlab.osti.gov/giridharnandipati/ksome}
 }
 
@@ -362,8 +348,6 @@
   author = {Deng, Zeyu},
   title = {{kMCpy: Kinetic Monte Carlo Simulation using Python}},
   year = {2022},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/caneparesearch/kMCpy}
 }
 
@@ -371,8 +355,6 @@
   author = {Alberdi, Joseba and Albi, Julen},
   title = {{Morphokinetics}},
   year = {2018},
-  publisher = {GitHub},
-  journal = {GitHub repository},
   url = {https://github.com/dipc-cc/Morphokinetics}
 }
 
