### 2024-06-11 sntioudis/papreca#22 @sntioudis
I removed the journal and publisher entries from all misc citations in the paper.bib file.
