### 2024-06-11 #6714 @srmnitc
@sntioudis Looks like `Heiber2020Sep` is defined twice in the bib file, could you please remove one? That should fix the issue
