### 2024-06-11 f2d24eb
Update paper.bib
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
