+++ paper/paper.bib
@@ -234,19 +234,6 @@
 	doi = {10.21105/joss.02307}
 }
 
-@article{Heiber2020Sep,
-	author = {Heiber, Michael C.},
-	title = {{Excimontec v1.0: An Open-Source Software Tool for Kinetic Monte Carlo Simulations of Organic Electronic Devices}},
-	journal = {Journal of Open Source Software},
-	volume = {5},
-	number = {53},
-	pages = {2307},
-	year = {2020},
-	month = sep,
-	issn = {2475-9066},
-	doi = {10.21105/joss.02307}
-}
-
 @article{Mitchell2023May,
 	author = {Mitchell, John A. and Abdeljawad, Fadi and Battaile, Corbett and Garcia-Cardona, Cristina and Holm, Elizabeth A. and Homer, Eric R. and Madison, Jon and Rodgers, Theron M. and Thompson, Aidan P. and Tikare, Veena and Webb, Ed and Plimpton, Steven J.},
 	title = {{Parallel simulation via SPPARKS of on-lattice kinetic and Metropolis Monte Carlo models for materials processing}},
