### 2024-06-12 #6714 @Kevin-Mattheus-Moerman
@sntioudis as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. All seems to be in order. I only have the below point that needs your attention:

- [ ] Check if `diffusions` should be `diffusion`.
