### 2024-06-12 #6714 @sntioudis
@Kevin-Mattheus-Moerman

>  Check if diffusions should be diffusion.

I made the relevant changes to the paper. For clarity, `diffusions` became `diffusion` and `depositions` became `deposition`.
