### 2024-06-13 8f4d906
Added JOSS citation to README.md
Signed-off-by: Stavros Ntioudis <42137665+sntioudis@users.noreply.github.com>
